% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_Aboriginal_Sex_TorresStraitIslander.R
\docType{data}
\name{LGA__Year_Aboriginal_Sex_TorresStraitIslander}
\alias{LGA__Year_Aboriginal_Sex_TorresStraitIslander}
\title{Aboriginal, Sex, TorresStraitIslander by LGA, Year}
\format{13,404 observations and 6 variables.}
\usage{
LGA__Year_Aboriginal_Sex_TorresStraitIslander
}
\description{
Number of personsAboriginal, Sex, TorresStraitIslander by LGA, Year
}
\keyword{datasets}
