% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_Aboriginal_TorresStraitIslander.R
\docType{data}
\name{LGA__Year_Aboriginal_TorresStraitIslander}
\alias{LGA__Year_Aboriginal_TorresStraitIslander}
\title{Aboriginal, TorresStraitIslander by LGA, Year}
\format{6,702 observations and 5 variables.}
\usage{
LGA__Year_Aboriginal_TorresStraitIslander
}
\description{
Number of personsAboriginal, TorresStraitIslander by LGA, Year
}
\keyword{datasets}
