% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_adults.R
\docType{data}
\name{LGA__Year_adults}
\alias{LGA__Year_adults}
\title{by LGA, Year}
\format{1,689 observations and 3 variables.}
\usage{
LGA__Year_adults
}
\description{
Number of adults by LGA, Year
}
\keyword{datasets}
