% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_Age.int.R
\docType{data}
\name{LGA__Year_Age.int}
\alias{LGA__Year_Age.int}
\title{Age.int by LGA, Year}
\format{136,755 observations and 4 variables.}
\usage{
LGA__Year_Age.int
}
\description{
Number of personsAge.int by LGA, Year
}
\keyword{datasets}
