% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_Age.int_Sex.R
\docType{data}
\name{LGA__Year_Age.int_Sex}
\alias{LGA__Year_Age.int_Sex}
\title{Age.int, Sex by LGA, Year}
\format{271,875 observations and 5 variables.}
\usage{
LGA__Year_Age.int_Sex
}
\description{
Number of personsAge.int, Sex by LGA, Year
}
\keyword{datasets}
