% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_Age.R
\docType{data}
\name{LGA__Year_Age}
\alias{LGA__Year_Age}
\title{Age by LGA, Year}
\format{18,579 observations and 4 variables.}
\usage{
LGA__Year_Age
}
\description{
Number of personsAge by LGA, Year
}
\keyword{datasets}
