% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_Age_adults.R
\docType{data}
\name{LGA__Year_Age_adults}
\alias{LGA__Year_Age_adults}
\title{Age, adults by LGA, Year}
\format{15,201 observations and 5 variables.}
\usage{
LGA__Year_Age_adults
}
\description{
VarAge, adults by LGA, Year #' @description IAge, adults by LGA, Year
}
\keyword{datasets}
