% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_Age_ChildrenEverBorn_females.R
\docType{data}
\name{LGA__Year_Age_ChildrenEverBorn_females}
\alias{LGA__Year_Age_ChildrenEverBorn_females}
\title{Age, ChildrenEverBorn by LGA, Year}
\format{202,680 observations and 5 variables.}
\usage{
LGA__Year_Age_ChildrenEverBorn_females
}
\description{
Number of femalesAge, ChildrenEverBorn by LGA, Year
}
\keyword{datasets}
