% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_Age_Denomination_Religion.R
\docType{data}
\name{LGA__Year_Age_Denomination_Religion}
\alias{LGA__Year_Age_Denomination_Religion}
\title{Age, Denomination, Religion by LGA, Year}
\format{425,628 observations and 6 variables.}
\usage{
LGA__Year_Age_Denomination_Religion
}
\description{
Number of personsAge, Denomination, Religion by LGA, Year
}
\keyword{datasets}
