% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_Age_FOE_adults.R
\docType{data}
\name{LGA__Year_Age_FOE_adults}
\alias{LGA__Year_Age_FOE_adults}
\title{Age, FOE by LGA, Year}
\format{197,613 observations and 5 variables.}
\usage{
LGA__Year_Age_FOE_adults
}
\description{
Number of adultsAge, FOE by LGA, Year
}
\keyword{datasets}
