% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_Age_LabourForceStatus_adults.R
\docType{data}
\name{LGA__Year_Age_LabourForceStatus_adults}
\alias{LGA__Year_Age_LabourForceStatus_adults}
\title{Age, LabourForceStatus by LGA, Year}
\format{175,656 observations and 5 variables.}
\usage{
LGA__Year_Age_LabourForceStatus_adults
}
\description{
Number of adultsAge, LabourForceStatus by LGA, Year
}
\keyword{datasets}
