% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_Age_LabourForceStatus_Sex_adults.R
\docType{data}
\name{LGA__Year_Age_LabourForceStatus_Sex_adults}
\alias{LGA__Year_Age_LabourForceStatus_Sex_adults}
\title{Age, LabourForceStatus, Sex by LGA, Year}
\format{351,312 observations and 6 variables.}
\usage{
LGA__Year_Age_LabourForceStatus_Sex_adults
}
\description{
Number of adultsAge, LabourForceStatus, Sex by LGA, Year
}
\keyword{datasets}
