% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_Age_MaritalStatus_Sex.R
\docType{data}
\name{LGA__Year_Age_MaritalStatus_Sex}
\alias{LGA__Year_Age_MaritalStatus_Sex}
\title{Age, MaritalStatus, Sex by LGA, Year}
\format{253,350 observations and 6 variables.}
\usage{
LGA__Year_Age_MaritalStatus_Sex
}
\description{
Number of personsAge, MaritalStatus, Sex by LGA, Year
}
\keyword{datasets}
