% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_Age_NonSchoolQualification_adults.R
\docType{data}
\name{LGA__Year_Age_NonSchoolQualification_adults}
\alias{LGA__Year_Age_NonSchoolQualification_adults}
\title{Age, NonSchoolQualification by LGA, Year}
\format{121,608 observations and 5 variables.}
\usage{
LGA__Year_Age_NonSchoolQualification_adults
}
\description{
Number of adultsAge, NonSchoolQualification by LGA, Year
}
\keyword{datasets}
