% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_Age_Sex.R
\docType{data}
\name{LGA__Year_Age_Sex}
\alias{LGA__Year_Age_Sex}
\title{Age, Sex by LGA, Year}
\format{37,158 observations and 5 variables.}
\usage{
LGA__Year_Age_Sex
}
\description{
Number of personsAge, Sex by LGA, Year
}
\keyword{datasets}
