% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_AustCitizen.R
\docType{data}
\name{LGA__Year_AustCitizen}
\alias{LGA__Year_AustCitizen}
\title{AustCitizen by LGA, Year}
\format{3,324 observations and 4 variables.}
\usage{
LGA__Year_AustCitizen
}
\description{
Number of personsAustCitizen by LGA, Year
}
\keyword{datasets}
