% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_AustCitizen_Sex.R
\docType{data}
\name{LGA__Year_AustCitizen_Sex}
\alias{LGA__Year_AustCitizen_Sex}
\title{AustCitizen, Sex by LGA, Year}
\format{5,013 observations and 5 variables.}
\usage{
LGA__Year_AustCitizen_Sex
}
\description{
Number of personsAustCitizen, Sex by LGA, Year
}
\keyword{datasets}
