% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_averageHouseholdSize.R
\docType{data}
\name{LGA__Year_averageHouseholdSize}
\alias{LGA__Year_averageHouseholdSize}
\title{by LGA, Year}
\format{1,689 observations and 3 variables.}
\usage{
LGA__Year_averageHouseholdSize
}
\description{
Average by LGA, Year #' @description Household by LGA, Year #' @description Size by LGA, Year
}
\keyword{datasets}
