% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_averageNumberOfPersonsPerBedroom.R
\docType{data}
\name{LGA__Year_averageNumberOfPersonsPerBedroom}
\alias{LGA__Year_averageNumberOfPersonsPerBedroom}
\title{by LGA, Year}
\format{1,689 observations and 3 variables.}
\usage{
LGA__Year_averageNumberOfPersonsPerBedroom
}
\description{
Average by LGA, Year #' @description Number by LGA, Year #' @description Of by LGA, Year #' @description Persons by LGA, Year #' @description Per by LGA, Year #' @description Bedroom by LGA, Year
}
\keyword{datasets}
