% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_BornAust.R
\docType{data}
\name{LGA__Year_BornAust}
\alias{LGA__Year_BornAust}
\title{BornAust by LGA, Year}
\format{3,378 observations and 4 variables.}
\usage{
LGA__Year_BornAust
}
\description{
Number of personsBornAust by LGA, Year
}
\keyword{datasets}
