% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_BornAust_Sex.R
\docType{data}
\name{LGA__Year_BornAust_Sex}
\alias{LGA__Year_BornAust_Sex}
\title{BornAust, Sex by LGA, Year}
\format{6,756 observations and 5 variables.}
\usage{
LGA__Year_BornAust_Sex
}
\description{
Number of personsBornAust, Sex by LGA, Year
}
\keyword{datasets}
