% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/LGA__Year_Children_CoupleFamily_TotalFamilyIncome.min_families_with_children.R
\docType{data}
\name{LGA__Year_Children_CoupleFamily_TotalFamilyIncome.min_families_with_children}
\alias{LGA__Year_Children_CoupleFamily_TotalFamilyIncome.min_families_with_children}
\title{Children, CoupleFamily, TotalFamilyIncome.min by LGA, Year}
\format{202,680 observations and 6 variables.}
\usage{
LGA__Year_Children_CoupleFamily_TotalFamilyIncome.min_families_with_children
}
\description{
Number of families_with_childrenChildren, CoupleFamily, TotalFamilyIncome.min by LGA, Year
}
\keyword{datasets}
