% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_CountryOfBirth.R
\docType{data}
\name{LGA__Year_CountryOfBirth}
\alias{LGA__Year_CountryOfBirth}
\title{CountryOfBirth by LGA, Year}
\format{62,493 observations and 4 variables.}
\usage{
LGA__Year_CountryOfBirth
}
\description{
Number of personsCountryOfBirth by LGA, Year
}
\keyword{datasets}
