% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_CountryOfBirth_Sex.R
\docType{data}
\name{LGA__Year_CountryOfBirth_Sex}
\alias{LGA__Year_CountryOfBirth_Sex}
\title{CountryOfBirth, Sex by LGA, Year}
\format{124,986 observations and 5 variables.}
\usage{
LGA__Year_CountryOfBirth_Sex
}
\description{
Number of personsCountryOfBirth, Sex by LGA, Year
}
\keyword{datasets}
