% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/LGA__Year_DwellingStructure_occupied_private_dwellings.R
\docType{data}
\name{LGA__Year_DwellingStructure_occupied_private_dwellings}
\alias{LGA__Year_DwellingStructure_occupied_private_dwellings}
\title{DwellingStructure by LGA, Year}
\format{8,445 observations and 4 variables.}
\usage{
LGA__Year_DwellingStructure_occupied_private_dwellings
}
\description{
Number of occupied_private_dwellingsDwellingStructure by LGA, Year
}
\keyword{datasets}
