% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/LGA__Year_DwellingStructure_TenureType_occupied_private_dwellings.R
\docType{data}
\name{LGA__Year_DwellingStructure_TenureType_occupied_private_dwellings}
\alias{LGA__Year_DwellingStructure_TenureType_occupied_private_dwellings}
\title{DwellingStructure, TenureType by LGA, Year}
\format{84,450 observations and 5 variables.}
\usage{
LGA__Year_DwellingStructure_TenureType_occupied_private_dwellings
}
\description{
Number of occupied_private_dwellingsDwellingStructure, TenureType by LGA, Year
}
\keyword{datasets}
