% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_EduInstitutionType_Sex_students.R
\docType{data}
\name{LGA__Year_EduInstitutionType_Sex_students}
\alias{LGA__Year_EduInstitutionType_Sex_students}
\title{EduInstitutionType, Sex by LGA, Year}
\format{23,646 observations and 5 variables.}
\usage{
LGA__Year_EduInstitutionType_Sex_students
}
\description{
Number of studentsEduInstitutionType, Sex by LGA, Year
}
\keyword{datasets}
