% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_EduInstitutionType_students.R
\docType{data}
\name{LGA__Year_EduInstitutionType_students}
\alias{LGA__Year_EduInstitutionType_students}
\title{EduInstitutionType by LGA, Year}
\format{11,823 observations and 4 variables.}
\usage{
LGA__Year_EduInstitutionType_students
}
\description{
Number of studentsEduInstitutionType by LGA, Year
}
\keyword{datasets}
