% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_employmentToPopulation.R
\docType{data}
\name{LGA__Year_employmentToPopulation}
\alias{LGA__Year_employmentToPopulation}
\title{by LGA, Year}
\format{1,689 observations and 3 variables.}
\usage{
LGA__Year_employmentToPopulation
}
\description{
Employment by LGA, Year #' @description To by LGA, Year #' @description Population by LGA, Year
}
\keyword{datasets}
