% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/LGA__Year_FamilyComposition_HasChild_HasDependantStudent_HasNonDependentChild_Rent.min_occupied_private_dwellings_being_rented.R
\docType{data}
\name{LGA__Year_FamilyComposition_HasChild_HasDependantStudent_HasNonDependentChild_Rent.min_occupied_private_dwellings_being_rented}
\alias{LGA__Year_FamilyComposition_HasChild_HasDependantStudent_HasNonDependentChild_Rent.min_occupied_private_dwellings_being_rented}
\title{FamilyComposition, HasChild, HasDependantStudent, HasNonDependentChild, Rent.min by LGA, Year}
\format{228,015 observations and 8 variables.}
\usage{

  LGA__Year_FamilyComposition_HasChild_HasDependantStudent_HasNonDependentChild_Rent.min_occupied_private_dwellings_being_rented
}
\description{
Number of occupied_private_dwellings_being_rentedFamilyComposition, HasChild, HasDependantStudent, HasNonDependentChild, Rent.min by LGA, Year
}
\keyword{datasets}
