% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/LGA__Year_FamilyComposition_HasChildren_TotalFamilyIncome.min_FullTimeWorkers_PartTimeWorkers_AwayFromWork_NotWorking_families.R
\docType{data}
\name{LGA__Year_FamilyComposition_HasChildren_TotalFamilyIncome.min_FullTimeWorkers_PartTimeWorkers_AwayFromWork_NotWorking_families}
\alias{LGA__Year_FamilyComposition_HasChildren_TotalFamilyIncome.min_FullTimeWorkers_PartTimeWorkers_AwayFromWork_NotWorking_families}
\title{FamilyComposition, HasChildren, TotalFamilyIncome.min, FullTimeWorkers, PartTimeWorkers, AwayFromWork, NotWorking by LGA, Year}
\format{658,710 observations and 10 variables.}
\usage{

  LGA__Year_FamilyComposition_HasChildren_TotalFamilyIncome.min_FullTimeWorkers_PartTimeWorkers_AwayFromWork_NotWorking_families
}
\description{
Number of familiesFamilyComposition, HasChildren, TotalFamilyIncome.min, FullTimeWorkers, PartTimeWorkers, AwayFromWork, NotWorking by LGA, Year
}
\keyword{datasets}
