% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_IndustryOfEmployment_Sex_workers.R
\docType{data}
\name{LGA__Year_IndustryOfEmployment_Sex_workers}
\alias{LGA__Year_IndustryOfEmployment_Sex_workers}
\title{IndustryOfEmployment, Sex by LGA, Year}
\format{67,560 observations and 5 variables.}
\usage{
LGA__Year_IndustryOfEmployment_Sex_workers
}
\description{
Number of workersIndustryOfEmployment, Sex by LGA, Year
}
\keyword{datasets}
