% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_IndustryOfEmployment_workers.R
\docType{data}
\name{LGA__Year_IndustryOfEmployment_workers}
\alias{LGA__Year_IndustryOfEmployment_workers}
\title{IndustryOfEmployment by LGA, Year}
\format{33,780 observations and 4 variables.}
\usage{
LGA__Year_IndustryOfEmployment_workers
}
\description{
Number of workersIndustryOfEmployment by LGA, Year
}
\keyword{datasets}
