% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_labourForceParticipation.R
\docType{data}
\name{LGA__Year_labourForceParticipation}
\alias{LGA__Year_labourForceParticipation}
\title{by LGA, Year}
\format{1,689 observations and 3 variables.}
\usage{
LGA__Year_labourForceParticipation
}
\description{
Labour by LGA, Year #' @description Force by LGA, Year #' @description Participation by LGA, Year
}
\keyword{datasets}
