% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_LabourForceStatus_adults.R
\docType{data}
\name{LGA__Year_LabourForceStatus_adults}
\alias{LGA__Year_LabourForceStatus_adults}
\title{LabourForceStatus by LGA, Year}
\format{10,134 observations and 4 variables.}
\usage{
LGA__Year_LabourForceStatus_adults
}
\description{
Number of adultsLabourForceStatus by LGA, Year
}
\keyword{datasets}
