% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_LabourForceStatus_Sex_adults.R
\docType{data}
\name{LGA__Year_LabourForceStatus_Sex_adults}
\alias{LGA__Year_LabourForceStatus_Sex_adults}
\title{LabourForceStatus, Sex by LGA, Year}
\format{20,268 observations and 5 variables.}
\usage{
LGA__Year_LabourForceStatus_Sex_adults
}
\description{
Number of adultsLabourForceStatus, Sex by LGA, Year
}
\keyword{datasets}
