% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/LGA__Year_Landlord_Rent.min_occupied_private_dwellings_being_rented.R
\docType{data}
\name{LGA__Year_Landlord_Rent.min_occupied_private_dwellings_being_rented}
\alias{LGA__Year_Landlord_Rent.min_occupied_private_dwellings_being_rented}
\title{Landlord, Rent.min by LGA, Year}
\format{152,010 observations and 5 variables.}
\usage{
LGA__Year_Landlord_Rent.min_occupied_private_dwellings_being_rented
}
\description{
Number of occupied_private_dwellings_being_rentedLandlord, Rent.min by LGA, Year
}
\keyword{datasets}
