% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_LanguageSpoken.R
\docType{data}
\name{LGA__Year_LanguageSpoken}
\alias{LGA__Year_LanguageSpoken}
\title{LanguageSpoken by LGA, Year}
\format{60,804 observations and 4 variables.}
\usage{
LGA__Year_LanguageSpoken
}
\description{
Number of personsLanguageSpoken by LGA, Year
}
\keyword{datasets}
