% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_LanguageSpoken_Sex.R
\docType{data}
\name{LGA__Year_LanguageSpoken_Sex}
\alias{LGA__Year_LanguageSpoken_Sex}
\title{LanguageSpoken, Sex by LGA, Year}
\format{121,608 observations and 5 variables.}
\usage{
LGA__Year_LanguageSpoken_Sex
}
\description{
Number of personsLanguageSpoken, Sex by LGA, Year
}
\keyword{datasets}
