% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_medianAgeOfPersons.R
\docType{data}
\name{LGA__Year_medianAgeOfPersons}
\alias{LGA__Year_medianAgeOfPersons}
\title{by LGA, Year}
\format{1,689 observations and 3 variables.}
\usage{
LGA__Year_medianAgeOfPersons
}
\description{
Median by LGA, Year #' @description Age by LGA, Year #' @description Of by LGA, Year #' @description Persons by LGA, Year
}
\keyword{datasets}
