% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_medianMortgageRepaymentMonthly.R
\docType{data}
\name{LGA__Year_medianMortgageRepaymentMonthly}
\alias{LGA__Year_medianMortgageRepaymentMonthly}
\title{by LGA, Year}
\format{1,689 observations and 3 variables.}
\usage{
LGA__Year_medianMortgageRepaymentMonthly
}
\description{
Median by LGA, Year #' @description Mortgage by LGA, Year #' @description Repayment by LGA, Year #' @description Monthly by LGA, Year
}
\keyword{datasets}
