% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_medianTotalHouseholdIncomeWeekly.R
\docType{data}
\name{LGA__Year_medianTotalHouseholdIncomeWeekly}
\alias{LGA__Year_medianTotalHouseholdIncomeWeekly}
\title{by LGA, Year}
\format{1,689 observations and 3 variables.}
\usage{
LGA__Year_medianTotalHouseholdIncomeWeekly
}
\description{
Median by LGA, Year #' @description Total by LGA, Year #' @description Household by LGA, Year #' @description Income by LGA, Year #' @description Weekly by LGA, Year
}
\keyword{datasets}
