% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_NonSchoolQualification.R
\docType{data}
\name{LGA__Year_NonSchoolQualification}
\alias{LGA__Year_NonSchoolQualification}
\title{NonSchoolQualification by LGA, Year}
\format{13,458 observations and 4 variables.}
\usage{
LGA__Year_NonSchoolQualification
}
\description{
Number of personsNonSchoolQualification by LGA, Year
}
\keyword{datasets}
