% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_NonSchoolQualification_Sex.R
\docType{data}
\name{LGA__Year_NonSchoolQualification_Sex}
\alias{LGA__Year_NonSchoolQualification_Sex}
\title{NonSchoolQualification, Sex by LGA, Year}
\format{26,916 observations and 5 variables.}
\usage{
LGA__Year_NonSchoolQualification_Sex
}
\description{
Number of personsNonSchoolQualification, Sex by LGA, Year
}
\keyword{datasets}
