% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_Occupation_Sex_workers.R
\docType{data}
\name{LGA__Year_Occupation_Sex_workers}
\alias{LGA__Year_Occupation_Sex_workers}
\title{Occupation, Sex by LGA, Year}
\format{30,402 observations and 5 variables.}
\usage{
LGA__Year_Occupation_Sex_workers
}
\description{
Number of workersOccupation, Sex by LGA, Year
}
\keyword{datasets}
