% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_Occupation_workers.R
\docType{data}
\name{LGA__Year_Occupation_workers}
\alias{LGA__Year_Occupation_workers}
\title{Occupation by LGA, Year}
\format{15,201 observations and 4 variables.}
\usage{
LGA__Year_Occupation_workers
}
\description{
Number of workersOccupation by LGA, Year
}
\keyword{datasets}
