% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_overseas_visitors.R
\docType{data}
\name{LGA__Year_overseas_visitors}
\alias{LGA__Year_overseas_visitors}
\title{by LGA, Year}
\format{1,689 observations and 3 variables.}
\usage{
LGA__Year_overseas_visitors
}
\description{
Number of overseas_visitors by LGA, Year
}
\keyword{datasets}
