% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_SameAddress1YearAgo.R
\docType{data}
\name{LGA__Year_SameAddress1YearAgo}
\alias{LGA__Year_SameAddress1YearAgo}
\title{SameAddress1YearAgo by LGA, Year}
\format{5,013 observations and 4 variables.}
\usage{
LGA__Year_SameAddress1YearAgo
}
\description{
Number of personsSameAddress1YearAgo by LGA, Year
}
\keyword{datasets}
