% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_SameAddress1YearAgo_Sex.R
\docType{data}
\name{LGA__Year_SameAddress1YearAgo_Sex}
\alias{LGA__Year_SameAddress1YearAgo_Sex}
\title{SameAddress1YearAgo, Sex by LGA, Year}
\format{10,026 observations and 5 variables.}
\usage{
LGA__Year_SameAddress1YearAgo_Sex
}
\description{
Number of personsSameAddress1YearAgo, Sex by LGA, Year
}
\keyword{datasets}
