% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_SameAddress5YearsAgo.R
\docType{data}
\name{LGA__Year_SameAddress5YearsAgo}
\alias{LGA__Year_SameAddress5YearsAgo}
\title{SameAddress5YearsAgo by LGA, Year}
\format{5,013 observations and 4 variables.}
\usage{
LGA__Year_SameAddress5YearsAgo
}
\description{
Number of personsSameAddress5YearsAgo by LGA, Year
}
\keyword{datasets}
