% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_SameAddress5YearsAgo_Sex.R
\docType{data}
\name{LGA__Year_SameAddress5YearsAgo_Sex}
\alias{LGA__Year_SameAddress5YearsAgo_Sex}
\title{SameAddress5YearsAgo, Sex by LGA, Year}
\format{10,026 observations and 5 variables.}
\usage{
LGA__Year_SameAddress5YearsAgo_Sex
}
\description{
Number of personsSameAddress5YearsAgo, Sex by LGA, Year
}
\keyword{datasets}
