% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_Sex.R
\docType{data}
\name{LGA__Year_Sex}
\alias{LGA__Year_Sex}
\title{Sex by LGA, Year}
\format{3,378 observations and 4 variables.}
\usage{
LGA__Year_Sex
}
\description{
Number of personsSex by LGA, Year
}
\keyword{datasets}
