% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_Sex_adults.R
\docType{data}
\name{LGA__Year_Sex_adults}
\alias{LGA__Year_Sex_adults}
\title{Sex by LGA, Year}
\format{3,378 observations and 4 variables.}
\usage{
LGA__Year_Sex_adults
}
\description{
Number of adultsSex by LGA, Year
}
\keyword{datasets}
