% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_Sex_overseas_visitors.R
\docType{data}
\name{LGA__Year_Sex_overseas_visitors}
\alias{LGA__Year_Sex_overseas_visitors}
\title{Sex by LGA, Year}
\format{3,378 observations and 4 variables.}
\usage{
LGA__Year_Sex_overseas_visitors
}
\description{
Number of overseas_visitorsSex by LGA, Year
}
\keyword{datasets}
