% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_Sex_SpeaksEnglishOnly.R
\docType{data}
\name{LGA__Year_Sex_SpeaksEnglishOnly}
\alias{LGA__Year_Sex_SpeaksEnglishOnly}
\title{Sex, SpeaksEnglishOnly by LGA, Year}
\format{6,756 observations and 5 variables.}
\usage{
LGA__Year_Sex_SpeaksEnglishOnly
}
\description{
Number of personsSex, SpeaksEnglishOnly by LGA, Year
}
\keyword{datasets}
