% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_SexemploymentToPopulation.R
\docType{data}
\name{LGA__Year_SexemploymentToPopulation}
\alias{LGA__Year_SexemploymentToPopulation}
\title{Sex by LGA, Year}
\format{3,378 observations and 4 variables.}
\usage{
LGA__Year_SexemploymentToPopulation
}
\description{
EmploymentSex by LGA, Year #' @description ToSex by LGA, Year #' @description PopulationSex by LGA, Year
}
\keyword{datasets}
