% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_SexlabourForceParticipation.R
\docType{data}
\name{LGA__Year_SexlabourForceParticipation}
\alias{LGA__Year_SexlabourForceParticipation}
\title{Sex by LGA, Year}
\format{3,378 observations and 4 variables.}
\usage{
LGA__Year_SexlabourForceParticipation
}
\description{
LabourSex by LGA, Year #' @description ForceSex by LGA, Year #' @description ParticipationSex by LGA, Year
}
\keyword{datasets}
