% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_Sexunemployment.R
\docType{data}
\name{LGA__Year_Sexunemployment}
\alias{LGA__Year_Sexunemployment}
\title{Sex by LGA, Year}
\format{3,378 observations and 4 variables.}
\usage{
LGA__Year_Sexunemployment
}
\description{
UnemploymentSex by LGA, Year
}
\keyword{datasets}
