% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_SpeaksEnglishOnly.R
\docType{data}
\name{LGA__Year_SpeaksEnglishOnly}
\alias{LGA__Year_SpeaksEnglishOnly}
\title{SpeaksEnglishOnly by LGA, Year}
\format{3,378 observations and 4 variables.}
\usage{
LGA__Year_SpeaksEnglishOnly
}
\description{
Number of personsSpeaksEnglishOnly by LGA, Year
}
\keyword{datasets}
