% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA__Year_unemployment.R
\docType{data}
\name{LGA__Year_unemployment}
\alias{LGA__Year_unemployment}
\title{by LGA, Year}
\format{1,689 observations and 3 variables.}
\usage{
LGA__Year_unemployment
}
\description{
Unemployment by LGA, Year
}
\keyword{datasets}
