% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_Aboriginal_Sex_TorresStraitIslander.R
\docType{data}
\name{SA2__Year_Aboriginal_Sex_TorresStraitIslander}
\alias{SA2__Year_Aboriginal_Sex_TorresStraitIslander}
\title{Aboriginal, Sex, TorresStraitIslander by SA2, Year}
\format{55,440 observations and 6 variables.}
\usage{
SA2__Year_Aboriginal_Sex_TorresStraitIslander
}
\description{
Number of personsAboriginal, Sex, TorresStraitIslander by SA2, Year
}
\keyword{datasets}
