% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_Aboriginal_TorresStraitIslander.R
\docType{data}
\name{SA2__Year_Aboriginal_TorresStraitIslander}
\alias{SA2__Year_Aboriginal_TorresStraitIslander}
\title{Aboriginal, TorresStraitIslander by SA2, Year}
\format{27,720 observations and 5 variables.}
\usage{
SA2__Year_Aboriginal_TorresStraitIslander
}
\description{
Number of personsAboriginal, TorresStraitIslander by SA2, Year
}
\keyword{datasets}
