% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_adults.R
\docType{data}
\name{SA2__Year_adults}
\alias{SA2__Year_adults}
\title{by SA2, Year}
\format{6,930 observations and 3 variables.}
\usage{
SA2__Year_adults
}
\description{
Number of adults by SA2, Year
}
\keyword{datasets}
