% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_Age.int.R
\docType{data}
\name{SA2__Year_Age.int}
\alias{SA2__Year_Age.int}
\title{Age.int by SA2, Year}
\format{561,330 observations and 4 variables.}
\usage{
SA2__Year_Age.int
}
\description{
Number of personsAge.int by SA2, Year
}
\keyword{datasets}
