% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_Age.int_Sex.R
\docType{data}
\name{SA2__Year_Age.int_Sex}
\alias{SA2__Year_Age.int_Sex}
\title{Age.int, Sex by SA2, Year}
\format{1,115,730 observations and 5 variables.}
\usage{
SA2__Year_Age.int_Sex
}
\description{
Number of personsAge.int, Sex by SA2, Year
}
\keyword{datasets}
