% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_Age.R
\docType{data}
\name{SA2__Year_Age}
\alias{SA2__Year_Age}
\title{Age by SA2, Year}
\format{76,230 observations and 4 variables.}
\usage{
SA2__Year_Age
}
\description{
Number of personsAge by SA2, Year
}
\keyword{datasets}
