% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_Age_adults.R
\docType{data}
\name{SA2__Year_Age_adults}
\alias{SA2__Year_Age_adults}
\title{Age, adults by SA2, Year}
\format{62,370 observations and 5 variables.}
\usage{
SA2__Year_Age_adults
}
\description{
VarAge, adults by SA2, Year #' @description IAge, adults by SA2, Year
}
\keyword{datasets}
