% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_Age_ChildrenEverBorn_females.R
\docType{data}
\name{SA2__Year_Age_ChildrenEverBorn_females}
\alias{SA2__Year_Age_ChildrenEverBorn_females}
\title{Age, ChildrenEverBorn by SA2, Year}
\format{831,600 observations and 5 variables.}
\usage{
SA2__Year_Age_ChildrenEverBorn_females
}
\description{
Number of femalesAge, ChildrenEverBorn by SA2, Year
}
\keyword{datasets}
