% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_Age_Denomination_Religion.R
\docType{data}
\name{SA2__Year_Age_Denomination_Religion}
\alias{SA2__Year_Age_Denomination_Religion}
\title{Age, Denomination, Religion by SA2, Year}
\format{1,746,360 observations and 6 variables.}
\usage{
SA2__Year_Age_Denomination_Religion
}
\description{
Number of personsAge, Denomination, Religion by SA2, Year
}
\keyword{datasets}
