% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_Age_FOE_adults.R
\docType{data}
\name{SA2__Year_Age_FOE_adults}
\alias{SA2__Year_Age_FOE_adults}
\title{Age, FOE by SA2, Year}
\format{810,810 observations and 5 variables.}
\usage{
SA2__Year_Age_FOE_adults
}
\description{
Number of adultsAge, FOE by SA2, Year
}
\keyword{datasets}
