% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_Age_LabourForceStatus_adults.R
\docType{data}
\name{SA2__Year_Age_LabourForceStatus_adults}
\alias{SA2__Year_Age_LabourForceStatus_adults}
\title{Age, LabourForceStatus by SA2, Year}
\format{720,720 observations and 5 variables.}
\usage{
SA2__Year_Age_LabourForceStatus_adults
}
\description{
Number of adultsAge, LabourForceStatus by SA2, Year
}
\keyword{datasets}
