% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_Age_LabourForceStatus_Sex_adults.R
\docType{data}
\name{SA2__Year_Age_LabourForceStatus_Sex_adults}
\alias{SA2__Year_Age_LabourForceStatus_Sex_adults}
\title{Age, LabourForceStatus, Sex by SA2, Year}
\format{1,441,440 observations and 6 variables.}
\usage{
SA2__Year_Age_LabourForceStatus_Sex_adults
}
\description{
Number of adultsAge, LabourForceStatus, Sex by SA2, Year
}
\keyword{datasets}
