% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_Age_MaritalStatus_Sex.R
\docType{data}
\name{SA2__Year_Age_MaritalStatus_Sex}
\alias{SA2__Year_Age_MaritalStatus_Sex}
\title{Age, MaritalStatus, Sex by SA2, Year}
\format{1,039,500 observations and 6 variables.}
\usage{
SA2__Year_Age_MaritalStatus_Sex
}
\description{
Number of personsAge, MaritalStatus, Sex by SA2, Year
}
\keyword{datasets}
