% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_Age_NonSchoolQualification_adults.R
\docType{data}
\name{SA2__Year_Age_NonSchoolQualification_adults}
\alias{SA2__Year_Age_NonSchoolQualification_adults}
\title{Age, NonSchoolQualification by SA2, Year}
\format{498,960 observations and 5 variables.}
\usage{
SA2__Year_Age_NonSchoolQualification_adults
}
\description{
Number of adultsAge, NonSchoolQualification by SA2, Year
}
\keyword{datasets}
