% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SA2__Year_Age_NonSchoolQualification_Sex_adults.R
\docType{data}
\name{SA2__Year_Age_NonSchoolQualification_Sex_adults}
\alias{SA2__Year_Age_NonSchoolQualification_Sex_adults}
\title{Age, NonSchoolQualification, Sex by SA2, Year}
\format{997,920 observations and 6 variables.}
\usage{
SA2__Year_Age_NonSchoolQualification_Sex_adults
}
\description{
Number of adultsAge, NonSchoolQualification, Sex by SA2, Year
}
\keyword{datasets}
