% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_Age_Sex.R
\docType{data}
\name{SA2__Year_Age_Sex}
\alias{SA2__Year_Age_Sex}
\title{Age, Sex by SA2, Year}
\format{152,460 observations and 5 variables.}
\usage{
SA2__Year_Age_Sex
}
\description{
Number of personsAge, Sex by SA2, Year
}
\keyword{datasets}
