% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_AustCitizen_Sex.R
\docType{data}
\name{SA2__Year_AustCitizen_Sex}
\alias{SA2__Year_AustCitizen_Sex}
\title{AustCitizen, Sex by SA2, Year}
\format{20,790 observations and 5 variables.}
\usage{
SA2__Year_AustCitizen_Sex
}
\description{
Number of personsAustCitizen, Sex by SA2, Year
}
\keyword{datasets}
