% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_averageHouseholdSize.R
\docType{data}
\name{SA2__Year_averageHouseholdSize}
\alias{SA2__Year_averageHouseholdSize}
\title{by SA2, Year}
\format{6,930 observations and 3 variables.}
\usage{
SA2__Year_averageHouseholdSize
}
\description{
Average by SA2, Year #' @description Household by SA2, Year #' @description Size by SA2, Year
}
\keyword{datasets}
