% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_averageNumberOfPersonsPerBedroom.R
\docType{data}
\name{SA2__Year_averageNumberOfPersonsPerBedroom}
\alias{SA2__Year_averageNumberOfPersonsPerBedroom}
\title{by SA2, Year}
\format{6,930 observations and 3 variables.}
\usage{
SA2__Year_averageNumberOfPersonsPerBedroom
}
\description{
Average by SA2, Year #' @description Number by SA2, Year #' @description Of by SA2, Year #' @description Persons by SA2, Year #' @description Per by SA2, Year #' @description Bedroom by SA2, Year
}
\keyword{datasets}
