% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_BornAust.R
\docType{data}
\name{SA2__Year_BornAust}
\alias{SA2__Year_BornAust}
\title{BornAust by SA2, Year}
\format{13,860 observations and 4 variables.}
\usage{
SA2__Year_BornAust
}
\description{
Number of personsBornAust by SA2, Year
}
\keyword{datasets}
