% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_BornAust_Sex.R
\docType{data}
\name{SA2__Year_BornAust_Sex}
\alias{SA2__Year_BornAust_Sex}
\title{BornAust, Sex by SA2, Year}
\format{27,720 observations and 5 variables.}
\usage{
SA2__Year_BornAust_Sex
}
\description{
Number of personsBornAust, Sex by SA2, Year
}
\keyword{datasets}
