% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SA2__Year_Children_CoupleFamily_TotalFamilyIncome.min_families_with_children.R
\docType{data}
\name{SA2__Year_Children_CoupleFamily_TotalFamilyIncome.min_families_with_children}
\alias{SA2__Year_Children_CoupleFamily_TotalFamilyIncome.min_families_with_children}
\title{Children, CoupleFamily, TotalFamilyIncome.min by SA2, Year}
\format{831,600 observations and 6 variables.}
\usage{
SA2__Year_Children_CoupleFamily_TotalFamilyIncome.min_families_with_children
}
\description{
Number of families_with_childrenChildren, CoupleFamily, TotalFamilyIncome.min by SA2, Year
}
\keyword{datasets}
