% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_CountryOfBirth.R
\docType{data}
\name{SA2__Year_CountryOfBirth}
\alias{SA2__Year_CountryOfBirth}
\title{CountryOfBirth by SA2, Year}
\format{256,410 observations and 4 variables.}
\usage{
SA2__Year_CountryOfBirth
}
\description{
Number of personsCountryOfBirth by SA2, Year
}
\keyword{datasets}
