% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_CountryOfBirth_Sex.R
\docType{data}
\name{SA2__Year_CountryOfBirth_Sex}
\alias{SA2__Year_CountryOfBirth_Sex}
\title{CountryOfBirth, Sex by SA2, Year}
\format{512,820 observations and 5 variables.}
\usage{
SA2__Year_CountryOfBirth_Sex
}
\description{
Number of personsCountryOfBirth, Sex by SA2, Year
}
\keyword{datasets}
