% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SA2__Year_DwellingStructure_occupied_private_dwellings.R
\docType{data}
\name{SA2__Year_DwellingStructure_occupied_private_dwellings}
\alias{SA2__Year_DwellingStructure_occupied_private_dwellings}
\title{DwellingStructure by SA2, Year}
\format{34,650 observations and 4 variables.}
\usage{
SA2__Year_DwellingStructure_occupied_private_dwellings
}
\description{
Number of occupied_private_dwellingsDwellingStructure by SA2, Year
}
\keyword{datasets}
