% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SA2__Year_DwellingStructure_UsualResidents.min_occupied_private_dwellings.R
\docType{data}
\name{SA2__Year_DwellingStructure_UsualResidents.min_occupied_private_dwellings}
\alias{SA2__Year_DwellingStructure_UsualResidents.min_occupied_private_dwellings}
\title{DwellingStructure, UsualResidents.min by SA2, Year}
\format{207,900 observations and 5 variables.}
\usage{
SA2__Year_DwellingStructure_UsualResidents.min_occupied_private_dwellings
}
\description{
Number of occupied_private_dwellingsDwellingStructure, UsualResidents.min by SA2, Year
}
\keyword{datasets}
