% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_EduInstitutionType_Sex_students.R
\docType{data}
\name{SA2__Year_EduInstitutionType_Sex_students}
\alias{SA2__Year_EduInstitutionType_Sex_students}
\title{EduInstitutionType, Sex by SA2, Year}
\format{97,020 observations and 5 variables.}
\usage{
SA2__Year_EduInstitutionType_Sex_students
}
\description{
Number of studentsEduInstitutionType, Sex by SA2, Year
}
\keyword{datasets}
