% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_EduInstitutionType_students.R
\docType{data}
\name{SA2__Year_EduInstitutionType_students}
\alias{SA2__Year_EduInstitutionType_students}
\title{EduInstitutionType by SA2, Year}
\format{48,510 observations and 4 variables.}
\usage{
SA2__Year_EduInstitutionType_students
}
\description{
Number of studentsEduInstitutionType by SA2, Year
}
\keyword{datasets}
