% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_employmentToPopulation.R
\docType{data}
\name{SA2__Year_employmentToPopulation}
\alias{SA2__Year_employmentToPopulation}
\title{by SA2, Year}
\format{6,930 observations and 3 variables.}
\usage{
SA2__Year_employmentToPopulation
}
\description{
Employment by SA2, Year #' @description To by SA2, Year #' @description Population by SA2, Year
}
\keyword{datasets}
