% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SA2__Year_FamilyComposition_HasChild_HasDependantStudent_HasNonDependentChild_Rent.min_occupied_private_dwellings_being_rented.R
\docType{data}
\name{SA2__Year_FamilyComposition_HasChild_HasDependantStudent_HasNonDependentChild_Rent.min_occupied_private_dwellings_being_rented}
\alias{SA2__Year_FamilyComposition_HasChild_HasDependantStudent_HasNonDependentChild_Rent.min_occupied_private_dwellings_being_rented}
\title{FamilyComposition, HasChild, HasDependantStudent, HasNonDependentChild, Rent.min by SA2, Year}
\format{935,550 observations and 8 variables.}
\usage{

  SA2__Year_FamilyComposition_HasChild_HasDependantStudent_HasNonDependentChild_Rent.min_occupied_private_dwellings_being_rented
}
\description{
Number of occupied_private_dwellings_being_rentedFamilyComposition, HasChild, HasDependantStudent, HasNonDependentChild, Rent.min by SA2, Year
}
\keyword{datasets}
