% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SA2__Year_FamilyComposition_HasChildren_TotalFamilyIncome.min_FullTimeWorkers_PartTimeWorkers_AwayFromWork_NotWorking_families.R
\docType{data}
\name{SA2__Year_FamilyComposition_HasChildren_TotalFamilyIncome.min_FullTimeWorkers_PartTimeWorkers_AwayFromWork_NotWorking_families}
\alias{SA2__Year_FamilyComposition_HasChildren_TotalFamilyIncome.min_FullTimeWorkers_PartTimeWorkers_AwayFromWork_NotWorking_families}
\title{FamilyComposition, HasChildren, TotalFamilyIncome.min, FullTimeWorkers, PartTimeWorkers, AwayFromWork, NotWorking by SA2, Year}
\format{2,702,700 observations and 10 variables.}
\usage{

  SA2__Year_FamilyComposition_HasChildren_TotalFamilyIncome.min_FullTimeWorkers_PartTimeWorkers_AwayFromWork_NotWorking_families
}
\description{
Number of familiesFamilyComposition, HasChildren, TotalFamilyIncome.min, FullTimeWorkers, PartTimeWorkers, AwayFromWork, NotWorking by SA2, Year
}
\keyword{datasets}
