% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_IndustryOfEmployment_workers.R
\docType{data}
\name{SA2__Year_IndustryOfEmployment_workers}
\alias{SA2__Year_IndustryOfEmployment_workers}
\title{IndustryOfEmployment by SA2, Year}
\format{138,600 observations and 4 variables.}
\usage{
SA2__Year_IndustryOfEmployment_workers
}
\description{
Number of workersIndustryOfEmployment by SA2, Year
}
\keyword{datasets}
