% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_labourForceParticipation.R
\docType{data}
\name{SA2__Year_labourForceParticipation}
\alias{SA2__Year_labourForceParticipation}
\title{by SA2, Year}
\format{6,930 observations and 3 variables.}
\usage{
SA2__Year_labourForceParticipation
}
\description{
Labour by SA2, Year #' @description Force by SA2, Year #' @description Participation by SA2, Year
}
\keyword{datasets}
