% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_LabourForceStatus_adults.R
\docType{data}
\name{SA2__Year_LabourForceStatus_adults}
\alias{SA2__Year_LabourForceStatus_adults}
\title{LabourForceStatus by SA2, Year}
\format{41,580 observations and 4 variables.}
\usage{
SA2__Year_LabourForceStatus_adults
}
\description{
Number of adultsLabourForceStatus by SA2, Year
}
\keyword{datasets}
