% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_LabourForceStatus_Sex_adults.R
\docType{data}
\name{SA2__Year_LabourForceStatus_Sex_adults}
\alias{SA2__Year_LabourForceStatus_Sex_adults}
\title{LabourForceStatus, Sex by SA2, Year}
\format{83,160 observations and 5 variables.}
\usage{
SA2__Year_LabourForceStatus_Sex_adults
}
\description{
Number of adultsLabourForceStatus, Sex by SA2, Year
}
\keyword{datasets}
