% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SA2__Year_Landlord_Rent.min_occupied_private_dwellings_being_rented.R
\docType{data}
\name{SA2__Year_Landlord_Rent.min_occupied_private_dwellings_being_rented}
\alias{SA2__Year_Landlord_Rent.min_occupied_private_dwellings_being_rented}
\title{Landlord, Rent.min by SA2, Year}
\format{623,700 observations and 5 variables.}
\usage{
SA2__Year_Landlord_Rent.min_occupied_private_dwellings_being_rented
}
\description{
Number of occupied_private_dwellings_being_rentedLandlord, Rent.min by SA2, Year
}
\keyword{datasets}
