% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_LanguageSpoken.R
\docType{data}
\name{SA2__Year_LanguageSpoken}
\alias{SA2__Year_LanguageSpoken}
\title{LanguageSpoken by SA2, Year}
\format{249,480 observations and 4 variables.}
\usage{
SA2__Year_LanguageSpoken
}
\description{
Number of personsLanguageSpoken by SA2, Year
}
\keyword{datasets}
