% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_LanguageSpoken_Sex.R
\docType{data}
\name{SA2__Year_LanguageSpoken_Sex}
\alias{SA2__Year_LanguageSpoken_Sex}
\title{LanguageSpoken, Sex by SA2, Year}
\format{498,960 observations and 5 variables.}
\usage{
SA2__Year_LanguageSpoken_Sex
}
\description{
Number of personsLanguageSpoken, Sex by SA2, Year
}
\keyword{datasets}
