% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_medianAgeOfPersons.R
\docType{data}
\name{SA2__Year_medianAgeOfPersons}
\alias{SA2__Year_medianAgeOfPersons}
\title{by SA2, Year}
\format{6,930 observations and 3 variables.}
\usage{
SA2__Year_medianAgeOfPersons
}
\description{
Median by SA2, Year #' @description Age by SA2, Year #' @description Of by SA2, Year #' @description Persons by SA2, Year
}
\keyword{datasets}
