% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_medianTotalHouseholdIncomeWeekly.R
\docType{data}
\name{SA2__Year_medianTotalHouseholdIncomeWeekly}
\alias{SA2__Year_medianTotalHouseholdIncomeWeekly}
\title{by SA2, Year}
\format{6,930 observations and 3 variables.}
\usage{
SA2__Year_medianTotalHouseholdIncomeWeekly
}
\description{
Median by SA2, Year #' @description Total by SA2, Year #' @description Household by SA2, Year #' @description Income by SA2, Year #' @description Weekly by SA2, Year
}
\keyword{datasets}
