% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_NonSchoolQualification.R
\docType{data}
\name{SA2__Year_NonSchoolQualification}
\alias{SA2__Year_NonSchoolQualification}
\title{NonSchoolQualification by SA2, Year}
\format{55,440 observations and 4 variables.}
\usage{
SA2__Year_NonSchoolQualification
}
\description{
Number of personsNonSchoolQualification by SA2, Year
}
\keyword{datasets}
