% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_NonSchoolQualification_Sex.R
\docType{data}
\name{SA2__Year_NonSchoolQualification_Sex}
\alias{SA2__Year_NonSchoolQualification_Sex}
\title{NonSchoolQualification, Sex by SA2, Year}
\format{110,880 observations and 5 variables.}
\usage{
SA2__Year_NonSchoolQualification_Sex
}
\description{
Number of personsNonSchoolQualification, Sex by SA2, Year
}
\keyword{datasets}
