% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_Occupation_Sex_workers.R
\docType{data}
\name{SA2__Year_Occupation_Sex_workers}
\alias{SA2__Year_Occupation_Sex_workers}
\title{Occupation, Sex by SA2, Year}
\format{124,740 observations and 5 variables.}
\usage{
SA2__Year_Occupation_Sex_workers
}
\description{
Number of workersOccupation, Sex by SA2, Year
}
\keyword{datasets}
