% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_Occupation_workers.R
\docType{data}
\name{SA2__Year_Occupation_workers}
\alias{SA2__Year_Occupation_workers}
\title{Occupation by SA2, Year}
\format{62,370 observations and 4 variables.}
\usage{
SA2__Year_Occupation_workers
}
\description{
Number of workersOccupation by SA2, Year
}
\keyword{datasets}
