% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SA2__Year_Rent.min_TotalHouseholdIncome.min_occupied_private_dwellings_being_rented.R
\docType{data}
\name{SA2__Year_Rent.min_TotalHouseholdIncome.min_occupied_private_dwellings_being_rented}
\alias{SA2__Year_Rent.min_TotalHouseholdIncome.min_occupied_private_dwellings_being_rented}
\title{Rent.min, TotalHouseholdIncome.min by SA2, Year}
\format{1,330,560 observations and 5 variables.}
\usage{

  SA2__Year_Rent.min_TotalHouseholdIncome.min_occupied_private_dwellings_being_rented
}
\description{
Number of occupied_private_dwellings_being_rentedRent.min, TotalHouseholdIncome.min by SA2, Year
}
\keyword{datasets}
