% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_SameAddress1YearAgo.R
\docType{data}
\name{SA2__Year_SameAddress1YearAgo}
\alias{SA2__Year_SameAddress1YearAgo}
\title{SameAddress1YearAgo by SA2, Year}
\format{20,790 observations and 4 variables.}
\usage{
SA2__Year_SameAddress1YearAgo
}
\description{
Number of personsSameAddress1YearAgo by SA2, Year
}
\keyword{datasets}
