% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_SameAddress1YearAgo_Sex.R
\docType{data}
\name{SA2__Year_SameAddress1YearAgo_Sex}
\alias{SA2__Year_SameAddress1YearAgo_Sex}
\title{SameAddress1YearAgo, Sex by SA2, Year}
\format{41,580 observations and 5 variables.}
\usage{
SA2__Year_SameAddress1YearAgo_Sex
}
\description{
Number of personsSameAddress1YearAgo, Sex by SA2, Year
}
\keyword{datasets}
