% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_SameAddress5YearsAgo.R
\docType{data}
\name{SA2__Year_SameAddress5YearsAgo}
\alias{SA2__Year_SameAddress5YearsAgo}
\title{SameAddress5YearsAgo by SA2, Year}
\format{20,790 observations and 4 variables.}
\usage{
SA2__Year_SameAddress5YearsAgo
}
\description{
Number of personsSameAddress5YearsAgo by SA2, Year
}
\keyword{datasets}
