% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_SameAddress5YearsAgo_Sex.R
\docType{data}
\name{SA2__Year_SameAddress5YearsAgo_Sex}
\alias{SA2__Year_SameAddress5YearsAgo_Sex}
\title{SameAddress5YearsAgo, Sex by SA2, Year}
\format{41,580 observations and 5 variables.}
\usage{
SA2__Year_SameAddress5YearsAgo_Sex
}
\description{
Number of personsSameAddress5YearsAgo, Sex by SA2, Year
}
\keyword{datasets}
