% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_Sex.R
\docType{data}
\name{SA2__Year_Sex}
\alias{SA2__Year_Sex}
\title{Sex by SA2, Year}
\format{13,860 observations and 4 variables.}
\usage{
SA2__Year_Sex
}
\description{
Number of personsSex by SA2, Year
}
\keyword{datasets}
