% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_Sex_adults.R
\docType{data}
\name{SA2__Year_Sex_adults}
\alias{SA2__Year_Sex_adults}
\title{Sex by SA2, Year}
\format{13,860 observations and 4 variables.}
\usage{
SA2__Year_Sex_adults
}
\description{
Number of adultsSex by SA2, Year
}
\keyword{datasets}
