% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_Sex_overseas_visitors.R
\docType{data}
\name{SA2__Year_Sex_overseas_visitors}
\alias{SA2__Year_Sex_overseas_visitors}
\title{Sex by SA2, Year}
\format{13,860 observations and 4 variables.}
\usage{
SA2__Year_Sex_overseas_visitors
}
\description{
Number of overseas_visitorsSex by SA2, Year
}
\keyword{datasets}
