% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_Sex_SpeaksEnglishOnly.R
\docType{data}
\name{SA2__Year_Sex_SpeaksEnglishOnly}
\alias{SA2__Year_Sex_SpeaksEnglishOnly}
\title{Sex, SpeaksEnglishOnly by SA2, Year}
\format{27,720 observations and 5 variables.}
\usage{
SA2__Year_Sex_SpeaksEnglishOnly
}
\description{
Number of personsSex, SpeaksEnglishOnly by SA2, Year
}
\keyword{datasets}
