% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_SexemploymentToPopulation.R
\docType{data}
\name{SA2__Year_SexemploymentToPopulation}
\alias{SA2__Year_SexemploymentToPopulation}
\title{Sex by SA2, Year}
\format{13,860 observations and 4 variables.}
\usage{
SA2__Year_SexemploymentToPopulation
}
\description{
EmploymentSex by SA2, Year #' @description ToSex by SA2, Year #' @description PopulationSex by SA2, Year
}
\keyword{datasets}
