% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_SexlabourForceParticipation.R
\docType{data}
\name{SA2__Year_SexlabourForceParticipation}
\alias{SA2__Year_SexlabourForceParticipation}
\title{Sex by SA2, Year}
\format{13,860 observations and 4 variables.}
\usage{
SA2__Year_SexlabourForceParticipation
}
\description{
LabourSex by SA2, Year #' @description ForceSex by SA2, Year #' @description ParticipationSex by SA2, Year
}
\keyword{datasets}
