% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_Sexunemployment.R
\docType{data}
\name{SA2__Year_Sexunemployment}
\alias{SA2__Year_Sexunemployment}
\title{Sex by SA2, Year}
\format{13,860 observations and 4 variables.}
\usage{
SA2__Year_Sexunemployment
}
\description{
UnemploymentSex by SA2, Year
}
\keyword{datasets}
