% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2__Year_SpeaksEnglishOnly.R
\docType{data}
\name{SA2__Year_SpeaksEnglishOnly}
\alias{SA2__Year_SpeaksEnglishOnly}
\title{SpeaksEnglishOnly by SA2, Year}
\format{13,860 observations and 4 variables.}
\usage{
SA2__Year_SpeaksEnglishOnly
}
\description{
Number of personsSpeaksEnglishOnly by SA2, Year
}
\keyword{datasets}
