% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_Aboriginal_Sex_TorresStraitIslander.R
\docType{data}
\name{SA3__Year_Aboriginal_Sex_TorresStraitIslander}
\alias{SA3__Year_Aboriginal_Sex_TorresStraitIslander}
\title{Aboriginal, Sex, TorresStraitIslander by SA3, Year}
\format{8,592 observations and 6 variables.}
\usage{
SA3__Year_Aboriginal_Sex_TorresStraitIslander
}
\description{
Number of personsAboriginal, Sex, TorresStraitIslander by SA3, Year
}
\keyword{datasets}
