% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_Aboriginal_TorresStraitIslander.R
\docType{data}
\name{SA3__Year_Aboriginal_TorresStraitIslander}
\alias{SA3__Year_Aboriginal_TorresStraitIslander}
\title{Aboriginal, TorresStraitIslander by SA3, Year}
\format{4,296 observations and 5 variables.}
\usage{
SA3__Year_Aboriginal_TorresStraitIslander
}
\description{
Number of personsAboriginal, TorresStraitIslander by SA3, Year
}
\keyword{datasets}
