% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_adults.R
\docType{data}
\name{SA3__Year_adults}
\alias{SA3__Year_adults}
\title{by SA3, Year}
\format{1,074 observations and 3 variables.}
\usage{
SA3__Year_adults
}
\description{
Number of adults by SA3, Year
}
\keyword{datasets}
