% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_Age.int.R
\docType{data}
\name{SA3__Year_Age.int}
\alias{SA3__Year_Age.int}
\title{Age.int by SA3, Year}
\format{86,994 observations and 4 variables.}
\usage{
SA3__Year_Age.int
}
\description{
Number of personsAge.int by SA3, Year
}
\keyword{datasets}
