% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_Age.int_Sex.R
\docType{data}
\name{SA3__Year_Age.int_Sex}
\alias{SA3__Year_Age.int_Sex}
\title{Age.int, Sex by SA3, Year}
\format{172,914 observations and 5 variables.}
\usage{
SA3__Year_Age.int_Sex
}
\description{
Number of personsAge.int, Sex by SA3, Year
}
\keyword{datasets}
