% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_Age.R
\docType{data}
\name{SA3__Year_Age}
\alias{SA3__Year_Age}
\title{Age by SA3, Year}
\format{11,814 observations and 4 variables.}
\usage{
SA3__Year_Age
}
\description{
Number of personsAge by SA3, Year
}
\keyword{datasets}
