% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_Age_adults.R
\docType{data}
\name{SA3__Year_Age_adults}
\alias{SA3__Year_Age_adults}
\title{Age, adults by SA3, Year}
\format{9,666 observations and 5 variables.}
\usage{
SA3__Year_Age_adults
}
\description{
VarAge, adults by SA3, Year #' @description IAge, adults by SA3, Year
}
\keyword{datasets}
