% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_Age_ChildrenEverBorn_females.R
\docType{data}
\name{SA3__Year_Age_ChildrenEverBorn_females}
\alias{SA3__Year_Age_ChildrenEverBorn_females}
\title{Age, ChildrenEverBorn by SA3, Year}
\format{128,880 observations and 5 variables.}
\usage{
SA3__Year_Age_ChildrenEverBorn_females
}
\description{
Number of femalesAge, ChildrenEverBorn by SA3, Year
}
\keyword{datasets}
