% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_Age_Denomination_Religion.R
\docType{data}
\name{SA3__Year_Age_Denomination_Religion}
\alias{SA3__Year_Age_Denomination_Religion}
\title{Age, Denomination, Religion by SA3, Year}
\format{270,648 observations and 6 variables.}
\usage{
SA3__Year_Age_Denomination_Religion
}
\description{
Number of personsAge, Denomination, Religion by SA3, Year
}
\keyword{datasets}
