% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_Age_FOE_adults.R
\docType{data}
\name{SA3__Year_Age_FOE_adults}
\alias{SA3__Year_Age_FOE_adults}
\title{Age, FOE by SA3, Year}
\format{125,658 observations and 5 variables.}
\usage{
SA3__Year_Age_FOE_adults
}
\description{
Number of adultsAge, FOE by SA3, Year
}
\keyword{datasets}
