% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_Age_LabourForceStatus_adults.R
\docType{data}
\name{SA3__Year_Age_LabourForceStatus_adults}
\alias{SA3__Year_Age_LabourForceStatus_adults}
\title{Age, LabourForceStatus by SA3, Year}
\format{111,696 observations and 5 variables.}
\usage{
SA3__Year_Age_LabourForceStatus_adults
}
\description{
Number of adultsAge, LabourForceStatus by SA3, Year
}
\keyword{datasets}
