% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_Age_LabourForceStatus_Sex_adults.R
\docType{data}
\name{SA3__Year_Age_LabourForceStatus_Sex_adults}
\alias{SA3__Year_Age_LabourForceStatus_Sex_adults}
\title{Age, LabourForceStatus, Sex by SA3, Year}
\format{223,392 observations and 6 variables.}
\usage{
SA3__Year_Age_LabourForceStatus_Sex_adults
}
\description{
Number of adultsAge, LabourForceStatus, Sex by SA3, Year
}
\keyword{datasets}
