% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_Age_MaritalStatus_Sex.R
\docType{data}
\name{SA3__Year_Age_MaritalStatus_Sex}
\alias{SA3__Year_Age_MaritalStatus_Sex}
\title{Age, MaritalStatus, Sex by SA3, Year}
\format{161,100 observations and 6 variables.}
\usage{
SA3__Year_Age_MaritalStatus_Sex
}
\description{
Number of personsAge, MaritalStatus, Sex by SA3, Year
}
\keyword{datasets}
