% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_Age_NonSchoolQualification_adults.R
\docType{data}
\name{SA3__Year_Age_NonSchoolQualification_adults}
\alias{SA3__Year_Age_NonSchoolQualification_adults}
\title{Age, NonSchoolQualification by SA3, Year}
\format{77,328 observations and 5 variables.}
\usage{
SA3__Year_Age_NonSchoolQualification_adults
}
\description{
Number of adultsAge, NonSchoolQualification by SA3, Year
}
\keyword{datasets}
