% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SA3__Year_Age_NonSchoolQualification_Sex_adults.R
\docType{data}
\name{SA3__Year_Age_NonSchoolQualification_Sex_adults}
\alias{SA3__Year_Age_NonSchoolQualification_Sex_adults}
\title{Age, NonSchoolQualification, Sex by SA3, Year}
\format{154,656 observations and 6 variables.}
\usage{
SA3__Year_Age_NonSchoolQualification_Sex_adults
}
\description{
Number of adultsAge, NonSchoolQualification, Sex by SA3, Year
}
\keyword{datasets}
