% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_Age_Sex.R
\docType{data}
\name{SA3__Year_Age_Sex}
\alias{SA3__Year_Age_Sex}
\title{Age, Sex by SA3, Year}
\format{23,628 observations and 5 variables.}
\usage{
SA3__Year_Age_Sex
}
\description{
Number of personsAge, Sex by SA3, Year
}
\keyword{datasets}
