% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_AustCitizen.R
\docType{data}
\name{SA3__Year_AustCitizen}
\alias{SA3__Year_AustCitizen}
\title{AustCitizen by SA3, Year}
\format{2,148 observations and 4 variables.}
\usage{
SA3__Year_AustCitizen
}
\description{
Number of personsAustCitizen by SA3, Year
}
\keyword{datasets}
