% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_AustCitizen_Sex.R
\docType{data}
\name{SA3__Year_AustCitizen_Sex}
\alias{SA3__Year_AustCitizen_Sex}
\title{AustCitizen, Sex by SA3, Year}
\format{3,222 observations and 5 variables.}
\usage{
SA3__Year_AustCitizen_Sex
}
\description{
Number of personsAustCitizen, Sex by SA3, Year
}
\keyword{datasets}
