% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_averageHouseholdSize.R
\docType{data}
\name{SA3__Year_averageHouseholdSize}
\alias{SA3__Year_averageHouseholdSize}
\title{by SA3, Year}
\format{1,074 observations and 3 variables.}
\usage{
SA3__Year_averageHouseholdSize
}
\description{
Average by SA3, Year #' @description Household by SA3, Year #' @description Size by SA3, Year
}
\keyword{datasets}
