% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_averageNumberOfPersonsPerBedroom.R
\docType{data}
\name{SA3__Year_averageNumberOfPersonsPerBedroom}
\alias{SA3__Year_averageNumberOfPersonsPerBedroom}
\title{by SA3, Year}
\format{1,074 observations and 3 variables.}
\usage{
SA3__Year_averageNumberOfPersonsPerBedroom
}
\description{
Average by SA3, Year #' @description Number by SA3, Year #' @description Of by SA3, Year #' @description Persons by SA3, Year #' @description Per by SA3, Year #' @description Bedroom by SA3, Year
}
\keyword{datasets}
