% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_BornAust.R
\docType{data}
\name{SA3__Year_BornAust}
\alias{SA3__Year_BornAust}
\title{BornAust by SA3, Year}
\format{2,148 observations and 4 variables.}
\usage{
SA3__Year_BornAust
}
\description{
Number of personsBornAust by SA3, Year
}
\keyword{datasets}
