% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_BornAust_Sex.R
\docType{data}
\name{SA3__Year_BornAust_Sex}
\alias{SA3__Year_BornAust_Sex}
\title{BornAust, Sex by SA3, Year}
\format{4,296 observations and 5 variables.}
\usage{
SA3__Year_BornAust_Sex
}
\description{
Number of personsBornAust, Sex by SA3, Year
}
\keyword{datasets}
