% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SA3__Year_Children_CoupleFamily_TotalFamilyIncome.min_families_with_children.R
\docType{data}
\name{SA3__Year_Children_CoupleFamily_TotalFamilyIncome.min_families_with_children}
\alias{SA3__Year_Children_CoupleFamily_TotalFamilyIncome.min_families_with_children}
\title{Children, CoupleFamily, TotalFamilyIncome.min by SA3, Year}
\format{128,880 observations and 6 variables.}
\usage{
SA3__Year_Children_CoupleFamily_TotalFamilyIncome.min_families_with_children
}
\description{
Number of families_with_childrenChildren, CoupleFamily, TotalFamilyIncome.min by SA3, Year
}
\keyword{datasets}
