% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_CountryOfBirth.R
\docType{data}
\name{SA3__Year_CountryOfBirth}
\alias{SA3__Year_CountryOfBirth}
\title{CountryOfBirth by SA3, Year}
\format{39,738 observations and 4 variables.}
\usage{
SA3__Year_CountryOfBirth
}
\description{
Number of personsCountryOfBirth by SA3, Year
}
\keyword{datasets}
