% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_CountryOfBirth_Sex.R
\docType{data}
\name{SA3__Year_CountryOfBirth_Sex}
\alias{SA3__Year_CountryOfBirth_Sex}
\title{CountryOfBirth, Sex by SA3, Year}
\format{79,476 observations and 5 variables.}
\usage{
SA3__Year_CountryOfBirth_Sex
}
\description{
Number of personsCountryOfBirth, Sex by SA3, Year
}
\keyword{datasets}
