% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SA3__Year_DwellingStructure_occupied_private_dwellings.R
\docType{data}
\name{SA3__Year_DwellingStructure_occupied_private_dwellings}
\alias{SA3__Year_DwellingStructure_occupied_private_dwellings}
\title{DwellingStructure by SA3, Year}
\format{5,370 observations and 4 variables.}
\usage{
SA3__Year_DwellingStructure_occupied_private_dwellings
}
\description{
Number of occupied_private_dwellingsDwellingStructure by SA3, Year
}
\keyword{datasets}
