% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SA3__Year_DwellingStructure_UsualResidents.min_occupied_private_dwellings.R
\docType{data}
\name{SA3__Year_DwellingStructure_UsualResidents.min_occupied_private_dwellings}
\alias{SA3__Year_DwellingStructure_UsualResidents.min_occupied_private_dwellings}
\title{DwellingStructure, UsualResidents.min by SA3, Year}
\format{32,220 observations and 5 variables.}
\usage{
SA3__Year_DwellingStructure_UsualResidents.min_occupied_private_dwellings
}
\description{
Number of occupied_private_dwellingsDwellingStructure, UsualResidents.min by SA3, Year
}
\keyword{datasets}
