% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_EduInstitutionType_Sex_students.R
\docType{data}
\name{SA3__Year_EduInstitutionType_Sex_students}
\alias{SA3__Year_EduInstitutionType_Sex_students}
\title{EduInstitutionType, Sex by SA3, Year}
\format{15,036 observations and 5 variables.}
\usage{
SA3__Year_EduInstitutionType_Sex_students
}
\description{
Number of studentsEduInstitutionType, Sex by SA3, Year
}
\keyword{datasets}
