% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_EduInstitutionType_students.R
\docType{data}
\name{SA3__Year_EduInstitutionType_students}
\alias{SA3__Year_EduInstitutionType_students}
\title{EduInstitutionType by SA3, Year}
\format{7,518 observations and 4 variables.}
\usage{
SA3__Year_EduInstitutionType_students
}
\description{
Number of studentsEduInstitutionType by SA3, Year
}
\keyword{datasets}
