% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_employmentToPopulation.R
\docType{data}
\name{SA3__Year_employmentToPopulation}
\alias{SA3__Year_employmentToPopulation}
\title{by SA3, Year}
\format{1,074 observations and 3 variables.}
\usage{
SA3__Year_employmentToPopulation
}
\description{
Employment by SA3, Year #' @description To by SA3, Year #' @description Population by SA3, Year
}
\keyword{datasets}
