% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SA3__Year_FamilyComposition_HasChild_HasDependantStudent_HasNonDependentChild_Rent.min_occupied_private_dwellings_being_rented.R
\docType{data}
\name{SA3__Year_FamilyComposition_HasChild_HasDependantStudent_HasNonDependentChild_Rent.min_occupied_private_dwellings_being_rented}
\alias{SA3__Year_FamilyComposition_HasChild_HasDependantStudent_HasNonDependentChild_Rent.min_occupied_private_dwellings_being_rented}
\title{FamilyComposition, HasChild, HasDependantStudent, HasNonDependentChild, Rent.min by SA3, Year}
\format{144,990 observations and 8 variables.}
\usage{

  SA3__Year_FamilyComposition_HasChild_HasDependantStudent_HasNonDependentChild_Rent.min_occupied_private_dwellings_being_rented
}
\description{
Number of occupied_private_dwellings_being_rentedFamilyComposition, HasChild, HasDependantStudent, HasNonDependentChild, Rent.min by SA3, Year
}
\keyword{datasets}
