% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SA3__Year_FamilyComposition_HasChildren_TotalFamilyIncome.min_FullTimeWorkers_PartTimeWorkers_AwayFromWork_NotWorking_families.R
\docType{data}
\name{SA3__Year_FamilyComposition_HasChildren_TotalFamilyIncome.min_FullTimeWorkers_PartTimeWorkers_AwayFromWork_NotWorking_families}
\alias{SA3__Year_FamilyComposition_HasChildren_TotalFamilyIncome.min_FullTimeWorkers_PartTimeWorkers_AwayFromWork_NotWorking_families}
\title{FamilyComposition, HasChildren, TotalFamilyIncome.min, FullTimeWorkers, PartTimeWorkers, AwayFromWork, NotWorking by SA3, Year}
\format{418,860 observations and 10 variables.}
\usage{

  SA3__Year_FamilyComposition_HasChildren_TotalFamilyIncome.min_FullTimeWorkers_PartTimeWorkers_AwayFromWork_NotWorking_families
}
\description{
Number of familiesFamilyComposition, HasChildren, TotalFamilyIncome.min, FullTimeWorkers, PartTimeWorkers, AwayFromWork, NotWorking by SA3, Year
}
\keyword{datasets}
