% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_IndustryOfEmployment_Sex_workers.R
\docType{data}
\name{SA3__Year_IndustryOfEmployment_Sex_workers}
\alias{SA3__Year_IndustryOfEmployment_Sex_workers}
\title{IndustryOfEmployment, Sex by SA3, Year}
\format{42,960 observations and 5 variables.}
\usage{
SA3__Year_IndustryOfEmployment_Sex_workers
}
\description{
Number of workersIndustryOfEmployment, Sex by SA3, Year
}
\keyword{datasets}
