% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_IndustryOfEmployment_workers.R
\docType{data}
\name{SA3__Year_IndustryOfEmployment_workers}
\alias{SA3__Year_IndustryOfEmployment_workers}
\title{IndustryOfEmployment by SA3, Year}
\format{21,480 observations and 4 variables.}
\usage{
SA3__Year_IndustryOfEmployment_workers
}
\description{
Number of workersIndustryOfEmployment by SA3, Year
}
\keyword{datasets}
