% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_labourForceParticipation.R
\docType{data}
\name{SA3__Year_labourForceParticipation}
\alias{SA3__Year_labourForceParticipation}
\title{by SA3, Year}
\format{1,074 observations and 3 variables.}
\usage{
SA3__Year_labourForceParticipation
}
\description{
Labour by SA3, Year #' @description Force by SA3, Year #' @description Participation by SA3, Year
}
\keyword{datasets}
