% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_LabourForceStatus_adults.R
\docType{data}
\name{SA3__Year_LabourForceStatus_adults}
\alias{SA3__Year_LabourForceStatus_adults}
\title{LabourForceStatus by SA3, Year}
\format{6,444 observations and 4 variables.}
\usage{
SA3__Year_LabourForceStatus_adults
}
\description{
Number of adultsLabourForceStatus by SA3, Year
}
\keyword{datasets}
