% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_LabourForceStatus_Sex_adults.R
\docType{data}
\name{SA3__Year_LabourForceStatus_Sex_adults}
\alias{SA3__Year_LabourForceStatus_Sex_adults}
\title{LabourForceStatus, Sex by SA3, Year}
\format{12,888 observations and 5 variables.}
\usage{
SA3__Year_LabourForceStatus_Sex_adults
}
\description{
Number of adultsLabourForceStatus, Sex by SA3, Year
}
\keyword{datasets}
