% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SA3__Year_Landlord_Rent.min_occupied_private_dwellings_being_rented.R
\docType{data}
\name{SA3__Year_Landlord_Rent.min_occupied_private_dwellings_being_rented}
\alias{SA3__Year_Landlord_Rent.min_occupied_private_dwellings_being_rented}
\title{Landlord, Rent.min by SA3, Year}
\format{96,660 observations and 5 variables.}
\usage{
SA3__Year_Landlord_Rent.min_occupied_private_dwellings_being_rented
}
\description{
Number of occupied_private_dwellings_being_rentedLandlord, Rent.min by SA3, Year
}
\keyword{datasets}
