% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_LanguageSpoken.R
\docType{data}
\name{SA3__Year_LanguageSpoken}
\alias{SA3__Year_LanguageSpoken}
\title{LanguageSpoken by SA3, Year}
\format{38,664 observations and 4 variables.}
\usage{
SA3__Year_LanguageSpoken
}
\description{
Number of personsLanguageSpoken by SA3, Year
}
\keyword{datasets}
