% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_LanguageSpoken_Sex.R
\docType{data}
\name{SA3__Year_LanguageSpoken_Sex}
\alias{SA3__Year_LanguageSpoken_Sex}
\title{LanguageSpoken, Sex by SA3, Year}
\format{77,328 observations and 5 variables.}
\usage{
SA3__Year_LanguageSpoken_Sex
}
\description{
Number of personsLanguageSpoken, Sex by SA3, Year
}
\keyword{datasets}
