% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_medianAgeOfPersons.R
\docType{data}
\name{SA3__Year_medianAgeOfPersons}
\alias{SA3__Year_medianAgeOfPersons}
\title{by SA3, Year}
\format{1,074 observations and 3 variables.}
\usage{
SA3__Year_medianAgeOfPersons
}
\description{
Median by SA3, Year #' @description Age by SA3, Year #' @description Of by SA3, Year #' @description Persons by SA3, Year
}
\keyword{datasets}
