% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_medianMortgageRepaymentMonthly.R
\docType{data}
\name{SA3__Year_medianMortgageRepaymentMonthly}
\alias{SA3__Year_medianMortgageRepaymentMonthly}
\title{by SA3, Year}
\format{1,074 observations and 3 variables.}
\usage{
SA3__Year_medianMortgageRepaymentMonthly
}
\description{
Median by SA3, Year #' @description Mortgage by SA3, Year #' @description Repayment by SA3, Year #' @description Monthly by SA3, Year
}
\keyword{datasets}
