% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_medianRentWeekly.R
\docType{data}
\name{SA3__Year_medianRentWeekly}
\alias{SA3__Year_medianRentWeekly}
\title{by SA3, Year}
\format{1,074 observations and 3 variables.}
\usage{
SA3__Year_medianRentWeekly
}
\description{
Median by SA3, Year #' @description Rent by SA3, Year #' @description Weekly by SA3, Year
}
\keyword{datasets}
