% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_medianTotalHouseholdIncomeWeekly.R
\docType{data}
\name{SA3__Year_medianTotalHouseholdIncomeWeekly}
\alias{SA3__Year_medianTotalHouseholdIncomeWeekly}
\title{by SA3, Year}
\format{1,074 observations and 3 variables.}
\usage{
SA3__Year_medianTotalHouseholdIncomeWeekly
}
\description{
Median by SA3, Year #' @description Total by SA3, Year #' @description Household by SA3, Year #' @description Income by SA3, Year #' @description Weekly by SA3, Year
}
\keyword{datasets}
