% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_NonSchoolQualification.R
\docType{data}
\name{SA3__Year_NonSchoolQualification}
\alias{SA3__Year_NonSchoolQualification}
\title{NonSchoolQualification by SA3, Year}
\format{8,592 observations and 4 variables.}
\usage{
SA3__Year_NonSchoolQualification
}
\description{
Number of personsNonSchoolQualification by SA3, Year
}
\keyword{datasets}
