% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_NonSchoolQualification_Sex.R
\docType{data}
\name{SA3__Year_NonSchoolQualification_Sex}
\alias{SA3__Year_NonSchoolQualification_Sex}
\title{NonSchoolQualification, Sex by SA3, Year}
\format{17,184 observations and 5 variables.}
\usage{
SA3__Year_NonSchoolQualification_Sex
}
\description{
Number of personsNonSchoolQualification, Sex by SA3, Year
}
\keyword{datasets}
