% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_Occupation_workers.R
\docType{data}
\name{SA3__Year_Occupation_workers}
\alias{SA3__Year_Occupation_workers}
\title{Occupation by SA3, Year}
\format{9,666 observations and 4 variables.}
\usage{
SA3__Year_Occupation_workers
}
\description{
Number of workersOccupation by SA3, Year
}
\keyword{datasets}
