% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_overseas_visitors.R
\docType{data}
\name{SA3__Year_overseas_visitors}
\alias{SA3__Year_overseas_visitors}
\title{by SA3, Year}
\format{1,074 observations and 3 variables.}
\usage{
SA3__Year_overseas_visitors
}
\description{
Number of overseas_visitors by SA3, Year
}
\keyword{datasets}
