% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SA3__Year_Rent.min_TotalHouseholdIncome.min_occupied_private_dwellings_being_rented.R
\docType{data}
\name{SA3__Year_Rent.min_TotalHouseholdIncome.min_occupied_private_dwellings_being_rented}
\alias{SA3__Year_Rent.min_TotalHouseholdIncome.min_occupied_private_dwellings_being_rented}
\title{Rent.min, TotalHouseholdIncome.min by SA3, Year}
\format{206,208 observations and 5 variables.}
\usage{

  SA3__Year_Rent.min_TotalHouseholdIncome.min_occupied_private_dwellings_being_rented
}
\description{
Number of occupied_private_dwellings_being_rentedRent.min, TotalHouseholdIncome.min by SA3, Year
}
\keyword{datasets}
