% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_SameAddress1YearAgo.R
\docType{data}
\name{SA3__Year_SameAddress1YearAgo}
\alias{SA3__Year_SameAddress1YearAgo}
\title{SameAddress1YearAgo by SA3, Year}
\format{3,222 observations and 4 variables.}
\usage{
SA3__Year_SameAddress1YearAgo
}
\description{
Number of personsSameAddress1YearAgo by SA3, Year
}
\keyword{datasets}
