% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_SameAddress5YearsAgo.R
\docType{data}
\name{SA3__Year_SameAddress5YearsAgo}
\alias{SA3__Year_SameAddress5YearsAgo}
\title{SameAddress5YearsAgo by SA3, Year}
\format{3,222 observations and 4 variables.}
\usage{
SA3__Year_SameAddress5YearsAgo
}
\description{
Number of personsSameAddress5YearsAgo by SA3, Year
}
\keyword{datasets}
