% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_SameAddress5YearsAgo_Sex.R
\docType{data}
\name{SA3__Year_SameAddress5YearsAgo_Sex}
\alias{SA3__Year_SameAddress5YearsAgo_Sex}
\title{SameAddress5YearsAgo, Sex by SA3, Year}
\format{6,444 observations and 5 variables.}
\usage{
SA3__Year_SameAddress5YearsAgo_Sex
}
\description{
Number of personsSameAddress5YearsAgo, Sex by SA3, Year
}
\keyword{datasets}
