% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_Sex.R
\docType{data}
\name{SA3__Year_Sex}
\alias{SA3__Year_Sex}
\title{Sex by SA3, Year}
\format{2,148 observations and 4 variables.}
\usage{
SA3__Year_Sex
}
\description{
Number of personsSex by SA3, Year
}
\keyword{datasets}
