% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_Sex_adults.R
\docType{data}
\name{SA3__Year_Sex_adults}
\alias{SA3__Year_Sex_adults}
\title{Sex by SA3, Year}
\format{2,148 observations and 4 variables.}
\usage{
SA3__Year_Sex_adults
}
\description{
Number of adultsSex by SA3, Year
}
\keyword{datasets}
