% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_Sex_overseas_visitors.R
\docType{data}
\name{SA3__Year_Sex_overseas_visitors}
\alias{SA3__Year_Sex_overseas_visitors}
\title{Sex by SA3, Year}
\format{2,148 observations and 4 variables.}
\usage{
SA3__Year_Sex_overseas_visitors
}
\description{
Number of overseas_visitorsSex by SA3, Year
}
\keyword{datasets}
