% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_Sex_SpeaksEnglishOnly.R
\docType{data}
\name{SA3__Year_Sex_SpeaksEnglishOnly}
\alias{SA3__Year_Sex_SpeaksEnglishOnly}
\title{Sex, SpeaksEnglishOnly by SA3, Year}
\format{4,296 observations and 5 variables.}
\usage{
SA3__Year_Sex_SpeaksEnglishOnly
}
\description{
Number of personsSex, SpeaksEnglishOnly by SA3, Year
}
\keyword{datasets}
