% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_SexemploymentToPopulation.R
\docType{data}
\name{SA3__Year_SexemploymentToPopulation}
\alias{SA3__Year_SexemploymentToPopulation}
\title{Sex by SA3, Year}
\format{2,148 observations and 4 variables.}
\usage{
SA3__Year_SexemploymentToPopulation
}
\description{
EmploymentSex by SA3, Year #' @description ToSex by SA3, Year #' @description PopulationSex by SA3, Year
}
\keyword{datasets}
