% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_SexlabourForceParticipation.R
\docType{data}
\name{SA3__Year_SexlabourForceParticipation}
\alias{SA3__Year_SexlabourForceParticipation}
\title{Sex by SA3, Year}
\format{2,148 observations and 4 variables.}
\usage{
SA3__Year_SexlabourForceParticipation
}
\description{
LabourSex by SA3, Year #' @description ForceSex by SA3, Year #' @description ParticipationSex by SA3, Year
}
\keyword{datasets}
