% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_Sexunemployment.R
\docType{data}
\name{SA3__Year_Sexunemployment}
\alias{SA3__Year_Sexunemployment}
\title{Sex by SA3, Year}
\format{2,148 observations and 4 variables.}
\usage{
SA3__Year_Sexunemployment
}
\description{
UnemploymentSex by SA3, Year
}
\keyword{datasets}
