% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_SpeaksEnglishOnly.R
\docType{data}
\name{SA3__Year_SpeaksEnglishOnly}
\alias{SA3__Year_SpeaksEnglishOnly}
\title{SpeaksEnglishOnly by SA3, Year}
\format{2,148 observations and 4 variables.}
\usage{
SA3__Year_SpeaksEnglishOnly
}
\description{
Number of personsSpeaksEnglishOnly by SA3, Year
}
\keyword{datasets}
