% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA3__Year_unemployment.R
\docType{data}
\name{SA3__Year_unemployment}
\alias{SA3__Year_unemployment}
\title{by SA3, Year}
\format{1,074 observations and 3 variables.}
\usage{
SA3__Year_unemployment
}
\description{
Unemployment by SA3, Year
}
\keyword{datasets}
