% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_Aboriginal_Sex_TorresStraitIslander.R
\docType{data}
\name{SA4__Year_Aboriginal_Sex_TorresStraitIslander}
\alias{SA4__Year_Aboriginal_Sex_TorresStraitIslander}
\title{Aboriginal, Sex, TorresStraitIslander by SA4, Year}
\format{2,568 observations and 6 variables.}
\usage{
SA4__Year_Aboriginal_Sex_TorresStraitIslander
}
\description{
Number of personsAboriginal, Sex, TorresStraitIslander by SA4, Year
}
\keyword{datasets}
