% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_Aboriginal_TorresStraitIslander.R
\docType{data}
\name{SA4__Year_Aboriginal_TorresStraitIslander}
\alias{SA4__Year_Aboriginal_TorresStraitIslander}
\title{Aboriginal, TorresStraitIslander by SA4, Year}
\format{1,284 observations and 5 variables.}
\usage{
SA4__Year_Aboriginal_TorresStraitIslander
}
\description{
Number of personsAboriginal, TorresStraitIslander by SA4, Year
}
\keyword{datasets}
