% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_adults.R
\docType{data}
\name{SA4__Year_adults}
\alias{SA4__Year_adults}
\title{by SA4, Year}
\format{321 observations and 3 variables.}
\usage{
SA4__Year_adults
}
\description{
Number of adults by SA4, Year
}
\keyword{datasets}
