% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_Age.int.R
\docType{data}
\name{SA4__Year_Age.int}
\alias{SA4__Year_Age.int}
\title{Age.int by SA4, Year}
\format{26,001 observations and 4 variables.}
\usage{
SA4__Year_Age.int
}
\description{
Number of personsAge.int by SA4, Year
}
\keyword{datasets}
