% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_Age.int_Sex.R
\docType{data}
\name{SA4__Year_Age.int_Sex}
\alias{SA4__Year_Age.int_Sex}
\title{Age.int, Sex by SA4, Year}
\format{51,681 observations and 5 variables.}
\usage{
SA4__Year_Age.int_Sex
}
\description{
Number of personsAge.int, Sex by SA4, Year
}
\keyword{datasets}
