% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_Age.R
\docType{data}
\name{SA4__Year_Age}
\alias{SA4__Year_Age}
\title{Age by SA4, Year}
\format{3,531 observations and 4 variables.}
\usage{
SA4__Year_Age
}
\description{
Number of personsAge by SA4, Year
}
\keyword{datasets}
