% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_Age_adults.R
\docType{data}
\name{SA4__Year_Age_adults}
\alias{SA4__Year_Age_adults}
\title{Age, adults by SA4, Year}
\format{2,889 observations and 5 variables.}
\usage{
SA4__Year_Age_adults
}
\description{
VarAge, adults by SA4, Year #' @description IAge, adults by SA4, Year
}
\keyword{datasets}
