% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_Age_ChildrenEverBorn_females.R
\docType{data}
\name{SA4__Year_Age_ChildrenEverBorn_females}
\alias{SA4__Year_Age_ChildrenEverBorn_females}
\title{Age, ChildrenEverBorn by SA4, Year}
\format{38,520 observations and 5 variables.}
\usage{
SA4__Year_Age_ChildrenEverBorn_females
}
\description{
Number of femalesAge, ChildrenEverBorn by SA4, Year
}
\keyword{datasets}
