% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_Age_Denomination_Religion.R
\docType{data}
\name{SA4__Year_Age_Denomination_Religion}
\alias{SA4__Year_Age_Denomination_Religion}
\title{Age, Denomination, Religion by SA4, Year}
\format{80,892 observations and 6 variables.}
\usage{
SA4__Year_Age_Denomination_Religion
}
\description{
Number of personsAge, Denomination, Religion by SA4, Year
}
\keyword{datasets}
