% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_Age_FOE_adults.R
\docType{data}
\name{SA4__Year_Age_FOE_adults}
\alias{SA4__Year_Age_FOE_adults}
\title{Age, FOE by SA4, Year}
\format{37,557 observations and 5 variables.}
\usage{
SA4__Year_Age_FOE_adults
}
\description{
Number of adultsAge, FOE by SA4, Year
}
\keyword{datasets}
