% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_Age_FOE_Sex_adults.R
\docType{data}
\name{SA4__Year_Age_FOE_Sex_adults}
\alias{SA4__Year_Age_FOE_Sex_adults}
\title{Age, FOE, Sex by SA4, Year}
\format{80,892 observations and 6 variables.}
\usage{
SA4__Year_Age_FOE_Sex_adults
}
\description{
Number of adultsAge, FOE, Sex by SA4, Year
}
\keyword{datasets}
