% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_Age_LabourForceStatus_adults.R
\docType{data}
\name{SA4__Year_Age_LabourForceStatus_adults}
\alias{SA4__Year_Age_LabourForceStatus_adults}
\title{Age, LabourForceStatus by SA4, Year}
\format{33,384 observations and 5 variables.}
\usage{
SA4__Year_Age_LabourForceStatus_adults
}
\description{
Number of adultsAge, LabourForceStatus by SA4, Year
}
\keyword{datasets}
