% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_Age_LabourForceStatus_Sex_adults.R
\docType{data}
\name{SA4__Year_Age_LabourForceStatus_Sex_adults}
\alias{SA4__Year_Age_LabourForceStatus_Sex_adults}
\title{Age, LabourForceStatus, Sex by SA4, Year}
\format{66,768 observations and 6 variables.}
\usage{
SA4__Year_Age_LabourForceStatus_Sex_adults
}
\description{
Number of adultsAge, LabourForceStatus, Sex by SA4, Year
}
\keyword{datasets}
