% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_Age_MaritalStatus_Sex.R
\docType{data}
\name{SA4__Year_Age_MaritalStatus_Sex}
\alias{SA4__Year_Age_MaritalStatus_Sex}
\title{Age, MaritalStatus, Sex by SA4, Year}
\format{48,150 observations and 6 variables.}
\usage{
SA4__Year_Age_MaritalStatus_Sex
}
\description{
Number of personsAge, MaritalStatus, Sex by SA4, Year
}
\keyword{datasets}
