% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_Age_NonSchoolQualification_adults.R
\docType{data}
\name{SA4__Year_Age_NonSchoolQualification_adults}
\alias{SA4__Year_Age_NonSchoolQualification_adults}
\title{Age, NonSchoolQualification by SA4, Year}
\format{23,112 observations and 5 variables.}
\usage{
SA4__Year_Age_NonSchoolQualification_adults
}
\description{
Number of adultsAge, NonSchoolQualification by SA4, Year
}
\keyword{datasets}
