% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_Age_Sex.R
\docType{data}
\name{SA4__Year_Age_Sex}
\alias{SA4__Year_Age_Sex}
\title{Age, Sex by SA4, Year}
\format{7,062 observations and 5 variables.}
\usage{
SA4__Year_Age_Sex
}
\description{
Number of personsAge, Sex by SA4, Year
}
\keyword{datasets}
