% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_AustCitizen.R
\docType{data}
\name{SA4__Year_AustCitizen}
\alias{SA4__Year_AustCitizen}
\title{AustCitizen by SA4, Year}
\format{642 observations and 4 variables.}
\usage{
SA4__Year_AustCitizen
}
\description{
Number of personsAustCitizen by SA4, Year
}
\keyword{datasets}
