% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_averageHouseholdSize.R
\docType{data}
\name{SA4__Year_averageHouseholdSize}
\alias{SA4__Year_averageHouseholdSize}
\title{by SA4, Year}
\format{321 observations and 3 variables.}
\usage{
SA4__Year_averageHouseholdSize
}
\description{
Average by SA4, Year #' @description Household by SA4, Year #' @description Size by SA4, Year
}
\keyword{datasets}
