% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_averageNumberOfPersonsPerBedroom.R
\docType{data}
\name{SA4__Year_averageNumberOfPersonsPerBedroom}
\alias{SA4__Year_averageNumberOfPersonsPerBedroom}
\title{by SA4, Year}
\format{321 observations and 3 variables.}
\usage{
SA4__Year_averageNumberOfPersonsPerBedroom
}
\description{
Average by SA4, Year #' @description Number by SA4, Year #' @description Of by SA4, Year #' @description Persons by SA4, Year #' @description Per by SA4, Year #' @description Bedroom by SA4, Year
}
\keyword{datasets}
