% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_BornAust.R
\docType{data}
\name{SA4__Year_BornAust}
\alias{SA4__Year_BornAust}
\title{BornAust by SA4, Year}
\format{642 observations and 4 variables.}
\usage{
SA4__Year_BornAust
}
\description{
Number of personsBornAust by SA4, Year
}
\keyword{datasets}
