% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_BornAust_Sex.R
\docType{data}
\name{SA4__Year_BornAust_Sex}
\alias{SA4__Year_BornAust_Sex}
\title{BornAust, Sex by SA4, Year}
\format{1,284 observations and 5 variables.}
\usage{
SA4__Year_BornAust_Sex
}
\description{
Number of personsBornAust, Sex by SA4, Year
}
\keyword{datasets}
