% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SA4__Year_Children_CoupleFamily_TotalFamilyIncome.min_families_with_children.R
\docType{data}
\name{SA4__Year_Children_CoupleFamily_TotalFamilyIncome.min_families_with_children}
\alias{SA4__Year_Children_CoupleFamily_TotalFamilyIncome.min_families_with_children}
\title{Children, CoupleFamily, TotalFamilyIncome.min by SA4, Year}
\format{38,520 observations and 6 variables.}
\usage{
SA4__Year_Children_CoupleFamily_TotalFamilyIncome.min_families_with_children
}
\description{
Number of families_with_childrenChildren, CoupleFamily, TotalFamilyIncome.min by SA4, Year
}
\keyword{datasets}
