% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_CountryOfBirth.R
\docType{data}
\name{SA4__Year_CountryOfBirth}
\alias{SA4__Year_CountryOfBirth}
\title{CountryOfBirth by SA4, Year}
\format{11,877 observations and 4 variables.}
\usage{
SA4__Year_CountryOfBirth
}
\description{
Number of personsCountryOfBirth by SA4, Year
}
\keyword{datasets}
