% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_CountryOfBirth_Sex.R
\docType{data}
\name{SA4__Year_CountryOfBirth_Sex}
\alias{SA4__Year_CountryOfBirth_Sex}
\title{CountryOfBirth, Sex by SA4, Year}
\format{23,754 observations and 5 variables.}
\usage{
SA4__Year_CountryOfBirth_Sex
}
\description{
Number of personsCountryOfBirth, Sex by SA4, Year
}
\keyword{datasets}
