% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SA4__Year_DwellingStructure_occupied_private_dwellings.R
\docType{data}
\name{SA4__Year_DwellingStructure_occupied_private_dwellings}
\alias{SA4__Year_DwellingStructure_occupied_private_dwellings}
\title{DwellingStructure by SA4, Year}
\format{1,605 observations and 4 variables.}
\usage{
SA4__Year_DwellingStructure_occupied_private_dwellings
}
\description{
Number of occupied_private_dwellingsDwellingStructure by SA4, Year
}
\keyword{datasets}
