% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SA4__Year_DwellingStructure_UsualResidents.min_occupied_private_dwellings.R
\docType{data}
\name{SA4__Year_DwellingStructure_UsualResidents.min_occupied_private_dwellings}
\alias{SA4__Year_DwellingStructure_UsualResidents.min_occupied_private_dwellings}
\title{DwellingStructure, UsualResidents.min by SA4, Year}
\format{9,630 observations and 5 variables.}
\usage{
SA4__Year_DwellingStructure_UsualResidents.min_occupied_private_dwellings
}
\description{
Number of occupied_private_dwellingsDwellingStructure, UsualResidents.min by SA4, Year
}
\keyword{datasets}
