% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_EduInstitutionType_Sex_students.R
\docType{data}
\name{SA4__Year_EduInstitutionType_Sex_students}
\alias{SA4__Year_EduInstitutionType_Sex_students}
\title{EduInstitutionType, Sex by SA4, Year}
\format{4,494 observations and 5 variables.}
\usage{
SA4__Year_EduInstitutionType_Sex_students
}
\description{
Number of studentsEduInstitutionType, Sex by SA4, Year
}
\keyword{datasets}
