% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_EduInstitutionType_students.R
\docType{data}
\name{SA4__Year_EduInstitutionType_students}
\alias{SA4__Year_EduInstitutionType_students}
\title{EduInstitutionType by SA4, Year}
\format{2,247 observations and 4 variables.}
\usage{
SA4__Year_EduInstitutionType_students
}
\description{
Number of studentsEduInstitutionType by SA4, Year
}
\keyword{datasets}
