% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SA4__Year_FamilyComposition_HasChild_HasDependantStudent_HasNonDependentChild_Rent.min_occupied_private_dwellings_being_rented.R
\docType{data}
\name{SA4__Year_FamilyComposition_HasChild_HasDependantStudent_HasNonDependentChild_Rent.min_occupied_private_dwellings_being_rented}
\alias{SA4__Year_FamilyComposition_HasChild_HasDependantStudent_HasNonDependentChild_Rent.min_occupied_private_dwellings_being_rented}
\title{FamilyComposition, HasChild, HasDependantStudent, HasNonDependentChild, Rent.min by SA4, Year}
\format{43,335 observations and 8 variables.}
\usage{

  SA4__Year_FamilyComposition_HasChild_HasDependantStudent_HasNonDependentChild_Rent.min_occupied_private_dwellings_being_rented
}
\description{
Number of occupied_private_dwellings_being_rentedFamilyComposition, HasChild, HasDependantStudent, HasNonDependentChild, Rent.min by SA4, Year
}
\keyword{datasets}
