% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SA4__Year_FamilyComposition_HasChildren_TotalFamilyIncome.min_FullTimeWorkers_PartTimeWorkers_AwayFromWork_NotWorking_families.R
\docType{data}
\name{SA4__Year_FamilyComposition_HasChildren_TotalFamilyIncome.min_FullTimeWorkers_PartTimeWorkers_AwayFromWork_NotWorking_families}
\alias{SA4__Year_FamilyComposition_HasChildren_TotalFamilyIncome.min_FullTimeWorkers_PartTimeWorkers_AwayFromWork_NotWorking_families}
\title{FamilyComposition, HasChildren, TotalFamilyIncome.min, FullTimeWorkers, PartTimeWorkers, AwayFromWork, NotWorking by SA4, Year}
\format{125,190 observations and 10 variables.}
\usage{

  SA4__Year_FamilyComposition_HasChildren_TotalFamilyIncome.min_FullTimeWorkers_PartTimeWorkers_AwayFromWork_NotWorking_families
}
\description{
Number of familiesFamilyComposition, HasChildren, TotalFamilyIncome.min, FullTimeWorkers, PartTimeWorkers, AwayFromWork, NotWorking by SA4, Year
}
\keyword{datasets}
