% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_IndustryOfEmployment_Sex_workers.R
\docType{data}
\name{SA4__Year_IndustryOfEmployment_Sex_workers}
\alias{SA4__Year_IndustryOfEmployment_Sex_workers}
\title{IndustryOfEmployment, Sex by SA4, Year}
\format{12,840 observations and 5 variables.}
\usage{
SA4__Year_IndustryOfEmployment_Sex_workers
}
\description{
Number of workersIndustryOfEmployment, Sex by SA4, Year
}
\keyword{datasets}
