% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_IndustryOfEmployment_workers.R
\docType{data}
\name{SA4__Year_IndustryOfEmployment_workers}
\alias{SA4__Year_IndustryOfEmployment_workers}
\title{IndustryOfEmployment by SA4, Year}
\format{6,420 observations and 4 variables.}
\usage{
SA4__Year_IndustryOfEmployment_workers
}
\description{
Number of workersIndustryOfEmployment by SA4, Year
}
\keyword{datasets}
