% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_labourForceParticipation.R
\docType{data}
\name{SA4__Year_labourForceParticipation}
\alias{SA4__Year_labourForceParticipation}
\title{by SA4, Year}
\format{321 observations and 3 variables.}
\usage{
SA4__Year_labourForceParticipation
}
\description{
Labour by SA4, Year #' @description Force by SA4, Year #' @description Participation by SA4, Year
}
\keyword{datasets}
