% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_LabourForceStatus_adults.R
\docType{data}
\name{SA4__Year_LabourForceStatus_adults}
\alias{SA4__Year_LabourForceStatus_adults}
\title{LabourForceStatus by SA4, Year}
\format{1,926 observations and 4 variables.}
\usage{
SA4__Year_LabourForceStatus_adults
}
\description{
Number of adultsLabourForceStatus by SA4, Year
}
\keyword{datasets}
