% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SA4__Year_Landlord_Rent.min_occupied_private_dwellings_being_rented.R
\docType{data}
\name{SA4__Year_Landlord_Rent.min_occupied_private_dwellings_being_rented}
\alias{SA4__Year_Landlord_Rent.min_occupied_private_dwellings_being_rented}
\title{Landlord, Rent.min by SA4, Year}
\format{28,890 observations and 5 variables.}
\usage{
SA4__Year_Landlord_Rent.min_occupied_private_dwellings_being_rented
}
\description{
Number of occupied_private_dwellings_being_rentedLandlord, Rent.min by SA4, Year
}
\keyword{datasets}
