% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_LanguageSpoken.R
\docType{data}
\name{SA4__Year_LanguageSpoken}
\alias{SA4__Year_LanguageSpoken}
\title{LanguageSpoken by SA4, Year}
\format{11,556 observations and 4 variables.}
\usage{
SA4__Year_LanguageSpoken
}
\description{
Number of personsLanguageSpoken by SA4, Year
}
\keyword{datasets}
