% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_LanguageSpoken_Sex.R
\docType{data}
\name{SA4__Year_LanguageSpoken_Sex}
\alias{SA4__Year_LanguageSpoken_Sex}
\title{LanguageSpoken, Sex by SA4, Year}
\format{23,112 observations and 5 variables.}
\usage{
SA4__Year_LanguageSpoken_Sex
}
\description{
Number of personsLanguageSpoken, Sex by SA4, Year
}
\keyword{datasets}
