% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_medianAgeOfPersons.R
\docType{data}
\name{SA4__Year_medianAgeOfPersons}
\alias{SA4__Year_medianAgeOfPersons}
\title{by SA4, Year}
\format{321 observations and 3 variables.}
\usage{
SA4__Year_medianAgeOfPersons
}
\description{
Median by SA4, Year #' @description Age by SA4, Year #' @description Of by SA4, Year #' @description Persons by SA4, Year
}
\keyword{datasets}
