% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_medianMortgageRepaymentMonthly.R
\docType{data}
\name{SA4__Year_medianMortgageRepaymentMonthly}
\alias{SA4__Year_medianMortgageRepaymentMonthly}
\title{by SA4, Year}
\format{321 observations and 3 variables.}
\usage{
SA4__Year_medianMortgageRepaymentMonthly
}
\description{
Median by SA4, Year #' @description Mortgage by SA4, Year #' @description Repayment by SA4, Year #' @description Monthly by SA4, Year
}
\keyword{datasets}
