% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_medianRentWeekly.R
\docType{data}
\name{SA4__Year_medianRentWeekly}
\alias{SA4__Year_medianRentWeekly}
\title{by SA4, Year}
\format{321 observations and 3 variables.}
\usage{
SA4__Year_medianRentWeekly
}
\description{
Median by SA4, Year #' @description Rent by SA4, Year #' @description Weekly by SA4, Year
}
\keyword{datasets}
