% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_medianTotalFamilyIncomeWeekly.R
\docType{data}
\name{SA4__Year_medianTotalFamilyIncomeWeekly}
\alias{SA4__Year_medianTotalFamilyIncomeWeekly}
\title{by SA4, Year}
\format{321 observations and 3 variables.}
\usage{
SA4__Year_medianTotalFamilyIncomeWeekly
}
\description{
Median by SA4, Year #' @description Total by SA4, Year #' @description Family by SA4, Year #' @description Income by SA4, Year #' @description Weekly by SA4, Year
}
\keyword{datasets}
