% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_NonSchoolQualification.R
\docType{data}
\name{SA4__Year_NonSchoolQualification}
\alias{SA4__Year_NonSchoolQualification}
\title{NonSchoolQualification by SA4, Year}
\format{2,568 observations and 4 variables.}
\usage{
SA4__Year_NonSchoolQualification
}
\description{
Number of personsNonSchoolQualification by SA4, Year
}
\keyword{datasets}
