% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_NonSchoolQualification_Sex.R
\docType{data}
\name{SA4__Year_NonSchoolQualification_Sex}
\alias{SA4__Year_NonSchoolQualification_Sex}
\title{NonSchoolQualification, Sex by SA4, Year}
\format{5,136 observations and 5 variables.}
\usage{
SA4__Year_NonSchoolQualification_Sex
}
\description{
Number of personsNonSchoolQualification, Sex by SA4, Year
}
\keyword{datasets}
