% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_Occupation_workers.R
\docType{data}
\name{SA4__Year_Occupation_workers}
\alias{SA4__Year_Occupation_workers}
\title{Occupation by SA4, Year}
\format{2,889 observations and 4 variables.}
\usage{
SA4__Year_Occupation_workers
}
\description{
Number of workersOccupation by SA4, Year
}
\keyword{datasets}
