% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_overseas_visitors.R
\docType{data}
\name{SA4__Year_overseas_visitors}
\alias{SA4__Year_overseas_visitors}
\title{by SA4, Year}
\format{321 observations and 3 variables.}
\usage{
SA4__Year_overseas_visitors
}
\description{
Number of overseas_visitors by SA4, Year
}
\keyword{datasets}
