% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_SameAddress1YearAgo.R
\docType{data}
\name{SA4__Year_SameAddress1YearAgo}
\alias{SA4__Year_SameAddress1YearAgo}
\title{SameAddress1YearAgo by SA4, Year}
\format{963 observations and 4 variables.}
\usage{
SA4__Year_SameAddress1YearAgo
}
\description{
Number of personsSameAddress1YearAgo by SA4, Year
}
\keyword{datasets}
