% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_SameAddress1YearAgo_Sex.R
\docType{data}
\name{SA4__Year_SameAddress1YearAgo_Sex}
\alias{SA4__Year_SameAddress1YearAgo_Sex}
\title{SameAddress1YearAgo, Sex by SA4, Year}
\format{1,926 observations and 5 variables.}
\usage{
SA4__Year_SameAddress1YearAgo_Sex
}
\description{
Number of personsSameAddress1YearAgo, Sex by SA4, Year
}
\keyword{datasets}
