% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_SameAddress5YearsAgo.R
\docType{data}
\name{SA4__Year_SameAddress5YearsAgo}
\alias{SA4__Year_SameAddress5YearsAgo}
\title{SameAddress5YearsAgo by SA4, Year}
\format{963 observations and 4 variables.}
\usage{
SA4__Year_SameAddress5YearsAgo
}
\description{
Number of personsSameAddress5YearsAgo by SA4, Year
}
\keyword{datasets}
