% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_SameAddress5YearsAgo_Sex.R
\docType{data}
\name{SA4__Year_SameAddress5YearsAgo_Sex}
\alias{SA4__Year_SameAddress5YearsAgo_Sex}
\title{SameAddress5YearsAgo, Sex by SA4, Year}
\format{1,926 observations and 5 variables.}
\usage{
SA4__Year_SameAddress5YearsAgo_Sex
}
\description{
Number of personsSameAddress5YearsAgo, Sex by SA4, Year
}
\keyword{datasets}
