% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_Sex.R
\docType{data}
\name{SA4__Year_Sex}
\alias{SA4__Year_Sex}
\title{Sex by SA4, Year}
\format{642 observations and 4 variables.}
\usage{
SA4__Year_Sex
}
\description{
Number of personsSex by SA4, Year
}
\keyword{datasets}
