% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_Sex_overseas_visitors.R
\docType{data}
\name{SA4__Year_Sex_overseas_visitors}
\alias{SA4__Year_Sex_overseas_visitors}
\title{Sex by SA4, Year}
\format{642 observations and 4 variables.}
\usage{
SA4__Year_Sex_overseas_visitors
}
\description{
Number of overseas_visitorsSex by SA4, Year
}
\keyword{datasets}
