% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_Sex_SpeaksEnglishOnly.R
\docType{data}
\name{SA4__Year_Sex_SpeaksEnglishOnly}
\alias{SA4__Year_Sex_SpeaksEnglishOnly}
\title{Sex, SpeaksEnglishOnly by SA4, Year}
\format{1,284 observations and 5 variables.}
\usage{
SA4__Year_Sex_SpeaksEnglishOnly
}
\description{
Number of personsSex, SpeaksEnglishOnly by SA4, Year
}
\keyword{datasets}
