% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_SexemploymentToPopulation.R
\docType{data}
\name{SA4__Year_SexemploymentToPopulation}
\alias{SA4__Year_SexemploymentToPopulation}
\title{Sex by SA4, Year}
\format{642 observations and 4 variables.}
\usage{
SA4__Year_SexemploymentToPopulation
}
\description{
EmploymentSex by SA4, Year #' @description ToSex by SA4, Year #' @description PopulationSex by SA4, Year
}
\keyword{datasets}
