% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_SexlabourForceParticipation.R
\docType{data}
\name{SA4__Year_SexlabourForceParticipation}
\alias{SA4__Year_SexlabourForceParticipation}
\title{Sex by SA4, Year}
\format{642 observations and 4 variables.}
\usage{
SA4__Year_SexlabourForceParticipation
}
\description{
LabourSex by SA4, Year #' @description ForceSex by SA4, Year #' @description ParticipationSex by SA4, Year
}
\keyword{datasets}
