% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_Sexunemployment.R
\docType{data}
\name{SA4__Year_Sexunemployment}
\alias{SA4__Year_Sexunemployment}
\title{Sex by SA4, Year}
\format{642 observations and 4 variables.}
\usage{
SA4__Year_Sexunemployment
}
\description{
UnemploymentSex by SA4, Year
}
\keyword{datasets}
