% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4__Year_SpeaksEnglishOnly.R
\docType{data}
\name{SA4__Year_SpeaksEnglishOnly}
\alias{SA4__Year_SpeaksEnglishOnly}
\title{SpeaksEnglishOnly by SA4, Year}
\format{642 observations and 4 variables.}
\usage{
SA4__Year_SpeaksEnglishOnly
}
\description{
Number of personsSpeaksEnglishOnly by SA4, Year
}
\keyword{datasets}
