% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_Aboriginal_Sex_TorresStraitIslander.R
\docType{data}
\name{STE__Year_Aboriginal_Sex_TorresStraitIslander}
\alias{STE__Year_Aboriginal_Sex_TorresStraitIslander}
\title{Aboriginal, Sex, TorresStraitIslander by STE, Year}
\format{216 observations and 6 variables.}
\usage{
STE__Year_Aboriginal_Sex_TorresStraitIslander
}
\description{
Number of personsAboriginal, Sex, TorresStraitIslander by STE, Year
}
\keyword{datasets}
