% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_Aboriginal_TorresStraitIslander.R
\docType{data}
\name{STE__Year_Aboriginal_TorresStraitIslander}
\alias{STE__Year_Aboriginal_TorresStraitIslander}
\title{Aboriginal, TorresStraitIslander by STE, Year}
\format{108 observations and 5 variables.}
\usage{
STE__Year_Aboriginal_TorresStraitIslander
}
\description{
Number of personsAboriginal, TorresStraitIslander by STE, Year
}
\keyword{datasets}
