% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_adults.R
\docType{data}
\name{STE__Year_adults}
\alias{STE__Year_adults}
\title{by STE, Year}
\format{27 observations and 3 variables.}
\usage{
STE__Year_adults
}
\description{
Number of adults by STE, Year
}
\keyword{datasets}
