% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_Age.int.R
\docType{data}
\name{STE__Year_Age.int}
\alias{STE__Year_Age.int}
\title{Age.int by STE, Year}
\format{2,187 observations and 4 variables.}
\usage{
STE__Year_Age.int
}
\description{
Number of personsAge.int by STE, Year
}
\keyword{datasets}
