% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_Age.int_Sex.R
\docType{data}
\name{STE__Year_Age.int_Sex}
\alias{STE__Year_Age.int_Sex}
\title{Age.int, Sex by STE, Year}
\format{4,347 observations and 5 variables.}
\usage{
STE__Year_Age.int_Sex
}
\description{
Number of personsAge.int, Sex by STE, Year
}
\keyword{datasets}
