% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_Age.R
\docType{data}
\name{STE__Year_Age}
\alias{STE__Year_Age}
\title{Age by STE, Year}
\format{297 observations and 4 variables.}
\usage{
STE__Year_Age
}
\description{
Number of personsAge by STE, Year
}
\keyword{datasets}
