% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_Age_adults.R
\docType{data}
\name{STE__Year_Age_adults}
\alias{STE__Year_Age_adults}
\title{Age, adults by STE, Year}
\format{243 observations and 5 variables.}
\usage{
STE__Year_Age_adults
}
\description{
VarAge, adults by STE, Year #' @description IAge, adults by STE, Year
}
\keyword{datasets}
