% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_Age_ChildrenEverBorn_females.R
\docType{data}
\name{STE__Year_Age_ChildrenEverBorn_females}
\alias{STE__Year_Age_ChildrenEverBorn_females}
\title{Age, ChildrenEverBorn by STE, Year}
\format{3,240 observations and 5 variables.}
\usage{
STE__Year_Age_ChildrenEverBorn_females
}
\description{
Number of femalesAge, ChildrenEverBorn by STE, Year
}
\keyword{datasets}
