% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_Age_Denomination_Religion.R
\docType{data}
\name{STE__Year_Age_Denomination_Religion}
\alias{STE__Year_Age_Denomination_Religion}
\title{Age, Denomination, Religion by STE, Year}
\format{6,804 observations and 6 variables.}
\usage{
STE__Year_Age_Denomination_Religion
}
\description{
Number of personsAge, Denomination, Religion by STE, Year
}
\keyword{datasets}
