% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_Age_FOE_adults.R
\docType{data}
\name{STE__Year_Age_FOE_adults}
\alias{STE__Year_Age_FOE_adults}
\title{Age, FOE by STE, Year}
\format{3,159 observations and 5 variables.}
\usage{
STE__Year_Age_FOE_adults
}
\description{
Number of adultsAge, FOE by STE, Year
}
\keyword{datasets}
