% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_Age_FOE_Sex_adults.R
\docType{data}
\name{STE__Year_Age_FOE_Sex_adults}
\alias{STE__Year_Age_FOE_Sex_adults}
\title{Age, FOE, Sex by STE, Year}
\format{6,804 observations and 6 variables.}
\usage{
STE__Year_Age_FOE_Sex_adults
}
\description{
Number of adultsAge, FOE, Sex by STE, Year
}
\keyword{datasets}
