% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_Age_LabourForceStatus_adults.R
\docType{data}
\name{STE__Year_Age_LabourForceStatus_adults}
\alias{STE__Year_Age_LabourForceStatus_adults}
\title{Age, LabourForceStatus by STE, Year}
\format{2,808 observations and 5 variables.}
\usage{
STE__Year_Age_LabourForceStatus_adults
}
\description{
Number of adultsAge, LabourForceStatus by STE, Year
}
\keyword{datasets}
