% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_Age_LabourForceStatus_Sex_adults.R
\docType{data}
\name{STE__Year_Age_LabourForceStatus_Sex_adults}
\alias{STE__Year_Age_LabourForceStatus_Sex_adults}
\title{Age, LabourForceStatus, Sex by STE, Year}
\format{5,616 observations and 6 variables.}
\usage{
STE__Year_Age_LabourForceStatus_Sex_adults
}
\description{
Number of adultsAge, LabourForceStatus, Sex by STE, Year
}
\keyword{datasets}
