% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_Age_MaritalStatus_Sex.R
\docType{data}
\name{STE__Year_Age_MaritalStatus_Sex}
\alias{STE__Year_Age_MaritalStatus_Sex}
\title{Age, MaritalStatus, Sex by STE, Year}
\format{4,050 observations and 6 variables.}
\usage{
STE__Year_Age_MaritalStatus_Sex
}
\description{
Number of personsAge, MaritalStatus, Sex by STE, Year
}
\keyword{datasets}
