% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_Age_NonSchoolQualification_adults.R
\docType{data}
\name{STE__Year_Age_NonSchoolQualification_adults}
\alias{STE__Year_Age_NonSchoolQualification_adults}
\title{Age, NonSchoolQualification by STE, Year}
\format{1,944 observations and 5 variables.}
\usage{
STE__Year_Age_NonSchoolQualification_adults
}
\description{
Number of adultsAge, NonSchoolQualification by STE, Year
}
\keyword{datasets}
