% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_Age_Sex.R
\docType{data}
\name{STE__Year_Age_Sex}
\alias{STE__Year_Age_Sex}
\title{Age, Sex by STE, Year}
\format{594 observations and 5 variables.}
\usage{
STE__Year_Age_Sex
}
\description{
Number of personsAge, Sex by STE, Year
}
\keyword{datasets}
