% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_AustCitizen.R
\docType{data}
\name{STE__Year_AustCitizen}
\alias{STE__Year_AustCitizen}
\title{AustCitizen by STE, Year}
\format{54 observations and 4 variables.}
\usage{
STE__Year_AustCitizen
}
\description{
Number of personsAustCitizen by STE, Year
}
\keyword{datasets}
