% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_AustCitizen_Sex.R
\docType{data}
\name{STE__Year_AustCitizen_Sex}
\alias{STE__Year_AustCitizen_Sex}
\title{AustCitizen, Sex by STE, Year}
\format{81 observations and 5 variables.}
\usage{
STE__Year_AustCitizen_Sex
}
\description{
Number of personsAustCitizen, Sex by STE, Year
}
\keyword{datasets}
