% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_averageHouseholdSize.R
\docType{data}
\name{STE__Year_averageHouseholdSize}
\alias{STE__Year_averageHouseholdSize}
\title{by STE, Year}
\format{27 observations and 3 variables.}
\usage{
STE__Year_averageHouseholdSize
}
\description{
Average by STE, Year #' @description Household by STE, Year #' @description Size by STE, Year
}
\keyword{datasets}
