% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_averageNumberOfPersonsPerBedroom.R
\docType{data}
\name{STE__Year_averageNumberOfPersonsPerBedroom}
\alias{STE__Year_averageNumberOfPersonsPerBedroom}
\title{by STE, Year}
\format{27 observations and 3 variables.}
\usage{
STE__Year_averageNumberOfPersonsPerBedroom
}
\description{
Average by STE, Year #' @description Number by STE, Year #' @description Of by STE, Year #' @description Persons by STE, Year #' @description Per by STE, Year #' @description Bedroom by STE, Year
}
\keyword{datasets}
