% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_BornAust.R
\docType{data}
\name{STE__Year_BornAust}
\alias{STE__Year_BornAust}
\title{BornAust by STE, Year}
\format{54 observations and 4 variables.}
\usage{
STE__Year_BornAust
}
\description{
Number of personsBornAust by STE, Year
}
\keyword{datasets}
