% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_BornAust_Sex.R
\docType{data}
\name{STE__Year_BornAust_Sex}
\alias{STE__Year_BornAust_Sex}
\title{BornAust, Sex by STE, Year}
\format{108 observations and 5 variables.}
\usage{
STE__Year_BornAust_Sex
}
\description{
Number of personsBornAust, Sex by STE, Year
}
\keyword{datasets}
