% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/STE__Year_Children_CoupleFamily_TotalFamilyIncome.min_families_with_children.R
\docType{data}
\name{STE__Year_Children_CoupleFamily_TotalFamilyIncome.min_families_with_children}
\alias{STE__Year_Children_CoupleFamily_TotalFamilyIncome.min_families_with_children}
\title{Children, CoupleFamily, TotalFamilyIncome.min by STE, Year}
\format{3,240 observations and 6 variables.}
\usage{
STE__Year_Children_CoupleFamily_TotalFamilyIncome.min_families_with_children
}
\description{
Number of families_with_childrenChildren, CoupleFamily, TotalFamilyIncome.min by STE, Year
}
\keyword{datasets}
