% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_CountryOfBirth.R
\docType{data}
\name{STE__Year_CountryOfBirth}
\alias{STE__Year_CountryOfBirth}
\title{CountryOfBirth by STE, Year}
\format{999 observations and 4 variables.}
\usage{
STE__Year_CountryOfBirth
}
\description{
Number of personsCountryOfBirth by STE, Year
}
\keyword{datasets}
