% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_CountryOfBirth_Sex.R
\docType{data}
\name{STE__Year_CountryOfBirth_Sex}
\alias{STE__Year_CountryOfBirth_Sex}
\title{CountryOfBirth, Sex by STE, Year}
\format{1,998 observations and 5 variables.}
\usage{
STE__Year_CountryOfBirth_Sex
}
\description{
Number of personsCountryOfBirth, Sex by STE, Year
}
\keyword{datasets}
