% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/STE__Year_DwellingStructure_occupied_private_dwellings.R
\docType{data}
\name{STE__Year_DwellingStructure_occupied_private_dwellings}
\alias{STE__Year_DwellingStructure_occupied_private_dwellings}
\title{DwellingStructure by STE, Year}
\format{135 observations and 4 variables.}
\usage{
STE__Year_DwellingStructure_occupied_private_dwellings
}
\description{
Number of occupied_private_dwellingsDwellingStructure by STE, Year
}
\keyword{datasets}
