% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/STE__Year_DwellingStructure_UsualResidents.min_occupied_private_dwellings.R
\docType{data}
\name{STE__Year_DwellingStructure_UsualResidents.min_occupied_private_dwellings}
\alias{STE__Year_DwellingStructure_UsualResidents.min_occupied_private_dwellings}
\title{DwellingStructure, UsualResidents.min by STE, Year}
\format{810 observations and 5 variables.}
\usage{
STE__Year_DwellingStructure_UsualResidents.min_occupied_private_dwellings
}
\description{
Number of occupied_private_dwellingsDwellingStructure, UsualResidents.min by STE, Year
}
\keyword{datasets}
