% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_EduInstitutionType_Sex_students.R
\docType{data}
\name{STE__Year_EduInstitutionType_Sex_students}
\alias{STE__Year_EduInstitutionType_Sex_students}
\title{EduInstitutionType, Sex by STE, Year}
\format{378 observations and 5 variables.}
\usage{
STE__Year_EduInstitutionType_Sex_students
}
\description{
Number of studentsEduInstitutionType, Sex by STE, Year
}
\keyword{datasets}
