% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_EduInstitutionType_students.R
\docType{data}
\name{STE__Year_EduInstitutionType_students}
\alias{STE__Year_EduInstitutionType_students}
\title{EduInstitutionType by STE, Year}
\format{189 observations and 4 variables.}
\usage{
STE__Year_EduInstitutionType_students
}
\description{
Number of studentsEduInstitutionType by STE, Year
}
\keyword{datasets}
