% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_employmentToPopulation.R
\docType{data}
\name{STE__Year_employmentToPopulation}
\alias{STE__Year_employmentToPopulation}
\title{by STE, Year}
\format{27 observations and 3 variables.}
\usage{
STE__Year_employmentToPopulation
}
\description{
Employment by STE, Year #' @description To by STE, Year #' @description Population by STE, Year
}
\keyword{datasets}
