% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/STE__Year_FamilyComposition_HasChild_HasDependantStudent_HasNonDependentChild_Rent.min_occupied_private_dwellings_being_rented.R
\docType{data}
\name{STE__Year_FamilyComposition_HasChild_HasDependantStudent_HasNonDependentChild_Rent.min_occupied_private_dwellings_being_rented}
\alias{STE__Year_FamilyComposition_HasChild_HasDependantStudent_HasNonDependentChild_Rent.min_occupied_private_dwellings_being_rented}
\title{FamilyComposition, HasChild, HasDependantStudent, HasNonDependentChild, Rent.min by STE, Year}
\format{3,645 observations and 8 variables.}
\usage{

  STE__Year_FamilyComposition_HasChild_HasDependantStudent_HasNonDependentChild_Rent.min_occupied_private_dwellings_being_rented
}
\description{
Number of occupied_private_dwellings_being_rentedFamilyComposition, HasChild, HasDependantStudent, HasNonDependentChild, Rent.min by STE, Year
}
\keyword{datasets}
