% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/STE__Year_FamilyComposition_HasChildren_TotalFamilyIncome.min_FullTimeWorkers_PartTimeWorkers_AwayFromWork_NotWorking_families.R
\docType{data}
\name{STE__Year_FamilyComposition_HasChildren_TotalFamilyIncome.min_FullTimeWorkers_PartTimeWorkers_AwayFromWork_NotWorking_families}
\alias{STE__Year_FamilyComposition_HasChildren_TotalFamilyIncome.min_FullTimeWorkers_PartTimeWorkers_AwayFromWork_NotWorking_families}
\title{FamilyComposition, HasChildren, TotalFamilyIncome.min, FullTimeWorkers, PartTimeWorkers, AwayFromWork, NotWorking by STE, Year}
\format{10,530 observations and 10 variables.}
\usage{

  STE__Year_FamilyComposition_HasChildren_TotalFamilyIncome.min_FullTimeWorkers_PartTimeWorkers_AwayFromWork_NotWorking_families
}
\description{
Number of familiesFamilyComposition, HasChildren, TotalFamilyIncome.min, FullTimeWorkers, PartTimeWorkers, AwayFromWork, NotWorking by STE, Year
}
\keyword{datasets}
