% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_IndustryOfEmployment_workers.R
\docType{data}
\name{STE__Year_IndustryOfEmployment_workers}
\alias{STE__Year_IndustryOfEmployment_workers}
\title{IndustryOfEmployment by STE, Year}
\format{540 observations and 4 variables.}
\usage{
STE__Year_IndustryOfEmployment_workers
}
\description{
Number of workersIndustryOfEmployment by STE, Year
}
\keyword{datasets}
