% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_labourForceParticipation.R
\docType{data}
\name{STE__Year_labourForceParticipation}
\alias{STE__Year_labourForceParticipation}
\title{by STE, Year}
\format{27 observations and 3 variables.}
\usage{
STE__Year_labourForceParticipation
}
\description{
Labour by STE, Year #' @description Force by STE, Year #' @description Participation by STE, Year
}
\keyword{datasets}
