% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_LabourForceStatus_adults.R
\docType{data}
\name{STE__Year_LabourForceStatus_adults}
\alias{STE__Year_LabourForceStatus_adults}
\title{LabourForceStatus by STE, Year}
\format{162 observations and 4 variables.}
\usage{
STE__Year_LabourForceStatus_adults
}
\description{
Number of adultsLabourForceStatus by STE, Year
}
\keyword{datasets}
