% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_LabourForceStatus_Sex_adults.R
\docType{data}
\name{STE__Year_LabourForceStatus_Sex_adults}
\alias{STE__Year_LabourForceStatus_Sex_adults}
\title{LabourForceStatus, Sex by STE, Year}
\format{324 observations and 5 variables.}
\usage{
STE__Year_LabourForceStatus_Sex_adults
}
\description{
Number of adultsLabourForceStatus, Sex by STE, Year
}
\keyword{datasets}
