% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/STE__Year_Landlord_Rent.min_occupied_private_dwellings_being_rented.R
\docType{data}
\name{STE__Year_Landlord_Rent.min_occupied_private_dwellings_being_rented}
\alias{STE__Year_Landlord_Rent.min_occupied_private_dwellings_being_rented}
\title{Landlord, Rent.min by STE, Year}
\format{2,430 observations and 5 variables.}
\usage{
STE__Year_Landlord_Rent.min_occupied_private_dwellings_being_rented
}
\description{
Number of occupied_private_dwellings_being_rentedLandlord, Rent.min by STE, Year
}
\keyword{datasets}
