% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_LanguageSpoken.R
\docType{data}
\name{STE__Year_LanguageSpoken}
\alias{STE__Year_LanguageSpoken}
\title{LanguageSpoken by STE, Year}
\format{972 observations and 4 variables.}
\usage{
STE__Year_LanguageSpoken
}
\description{
Number of personsLanguageSpoken by STE, Year
}
\keyword{datasets}
