% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_LanguageSpoken_Sex.R
\docType{data}
\name{STE__Year_LanguageSpoken_Sex}
\alias{STE__Year_LanguageSpoken_Sex}
\title{LanguageSpoken, Sex by STE, Year}
\format{1,944 observations and 5 variables.}
\usage{
STE__Year_LanguageSpoken_Sex
}
\description{
Number of personsLanguageSpoken, Sex by STE, Year
}
\keyword{datasets}
