% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_medianAgeOfPersons.R
\docType{data}
\name{STE__Year_medianAgeOfPersons}
\alias{STE__Year_medianAgeOfPersons}
\title{by STE, Year}
\format{27 observations and 3 variables.}
\usage{
STE__Year_medianAgeOfPersons
}
\description{
Median by STE, Year #' @description Age by STE, Year #' @description Of by STE, Year #' @description Persons by STE, Year
}
\keyword{datasets}
