% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_medianMortgageRepaymentMonthly.R
\docType{data}
\name{STE__Year_medianMortgageRepaymentMonthly}
\alias{STE__Year_medianMortgageRepaymentMonthly}
\title{by STE, Year}
\format{27 observations and 3 variables.}
\usage{
STE__Year_medianMortgageRepaymentMonthly
}
\description{
Median by STE, Year #' @description Mortgage by STE, Year #' @description Repayment by STE, Year #' @description Monthly by STE, Year
}
\keyword{datasets}
