% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_medianTotalHouseholdIncomeWeekly.R
\docType{data}
\name{STE__Year_medianTotalHouseholdIncomeWeekly}
\alias{STE__Year_medianTotalHouseholdIncomeWeekly}
\title{by STE, Year}
\format{27 observations and 3 variables.}
\usage{
STE__Year_medianTotalHouseholdIncomeWeekly
}
\description{
Median by STE, Year #' @description Total by STE, Year #' @description Household by STE, Year #' @description Income by STE, Year #' @description Weekly by STE, Year
}
\keyword{datasets}
