% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_NonSchoolQualification.R
\docType{data}
\name{STE__Year_NonSchoolQualification}
\alias{STE__Year_NonSchoolQualification}
\title{NonSchoolQualification by STE, Year}
\format{216 observations and 4 variables.}
\usage{
STE__Year_NonSchoolQualification
}
\description{
Number of personsNonSchoolQualification by STE, Year
}
\keyword{datasets}
