% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_NonSchoolQualification_Sex.R
\docType{data}
\name{STE__Year_NonSchoolQualification_Sex}
\alias{STE__Year_NonSchoolQualification_Sex}
\title{NonSchoolQualification, Sex by STE, Year}
\format{432 observations and 5 variables.}
\usage{
STE__Year_NonSchoolQualification_Sex
}
\description{
Number of personsNonSchoolQualification, Sex by STE, Year
}
\keyword{datasets}
