% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_Occupation_Sex_workers.R
\docType{data}
\name{STE__Year_Occupation_Sex_workers}
\alias{STE__Year_Occupation_Sex_workers}
\title{Occupation, Sex by STE, Year}
\format{486 observations and 5 variables.}
\usage{
STE__Year_Occupation_Sex_workers
}
\description{
Number of workersOccupation, Sex by STE, Year
}
\keyword{datasets}
