% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_Occupation_workers.R
\docType{data}
\name{STE__Year_Occupation_workers}
\alias{STE__Year_Occupation_workers}
\title{Occupation by STE, Year}
\format{243 observations and 4 variables.}
\usage{
STE__Year_Occupation_workers
}
\description{
Number of workersOccupation by STE, Year
}
\keyword{datasets}
