% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_overseas_visitors.R
\docType{data}
\name{STE__Year_overseas_visitors}
\alias{STE__Year_overseas_visitors}
\title{by STE, Year}
\format{27 observations and 3 variables.}
\usage{
STE__Year_overseas_visitors
}
\description{
Number of overseas_visitors by STE, Year
}
\keyword{datasets}
