% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/STE__Year_Rent.min_TotalHouseholdIncome.min_occupied_private_dwellings_being_rented.R
\docType{data}
\name{STE__Year_Rent.min_TotalHouseholdIncome.min_occupied_private_dwellings_being_rented}
\alias{STE__Year_Rent.min_TotalHouseholdIncome.min_occupied_private_dwellings_being_rented}
\title{Rent.min, TotalHouseholdIncome.min by STE, Year}
\format{5,184 observations and 5 variables.}
\usage{

  STE__Year_Rent.min_TotalHouseholdIncome.min_occupied_private_dwellings_being_rented
}
\description{
Number of occupied_private_dwellings_being_rentedRent.min, TotalHouseholdIncome.min by STE, Year
}
\keyword{datasets}
