% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_SameAddress1YearAgo.R
\docType{data}
\name{STE__Year_SameAddress1YearAgo}
\alias{STE__Year_SameAddress1YearAgo}
\title{SameAddress1YearAgo by STE, Year}
\format{81 observations and 4 variables.}
\usage{
STE__Year_SameAddress1YearAgo
}
\description{
Number of personsSameAddress1YearAgo by STE, Year
}
\keyword{datasets}
