% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_SameAddress1YearAgo_Sex.R
\docType{data}
\name{STE__Year_SameAddress1YearAgo_Sex}
\alias{STE__Year_SameAddress1YearAgo_Sex}
\title{SameAddress1YearAgo, Sex by STE, Year}
\format{162 observations and 5 variables.}
\usage{
STE__Year_SameAddress1YearAgo_Sex
}
\description{
Number of personsSameAddress1YearAgo, Sex by STE, Year
}
\keyword{datasets}
