% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_SameAddress5YearsAgo.R
\docType{data}
\name{STE__Year_SameAddress5YearsAgo}
\alias{STE__Year_SameAddress5YearsAgo}
\title{SameAddress5YearsAgo by STE, Year}
\format{81 observations and 4 variables.}
\usage{
STE__Year_SameAddress5YearsAgo
}
\description{
Number of personsSameAddress5YearsAgo by STE, Year
}
\keyword{datasets}
