% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_SameAddress5YearsAgo_Sex.R
\docType{data}
\name{STE__Year_SameAddress5YearsAgo_Sex}
\alias{STE__Year_SameAddress5YearsAgo_Sex}
\title{SameAddress5YearsAgo, Sex by STE, Year}
\format{162 observations and 5 variables.}
\usage{
STE__Year_SameAddress5YearsAgo_Sex
}
\description{
Number of personsSameAddress5YearsAgo, Sex by STE, Year
}
\keyword{datasets}
