% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_Sex.R
\docType{data}
\name{STE__Year_Sex}
\alias{STE__Year_Sex}
\title{Sex by STE, Year}
\format{54 observations and 4 variables.}
\usage{
STE__Year_Sex
}
\description{
Number of personsSex by STE, Year
}
\keyword{datasets}
