% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_Sex_adults.R
\docType{data}
\name{STE__Year_Sex_adults}
\alias{STE__Year_Sex_adults}
\title{Sex by STE, Year}
\format{54 observations and 4 variables.}
\usage{
STE__Year_Sex_adults
}
\description{
Number of adultsSex by STE, Year
}
\keyword{datasets}
