% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_Sex_overseas_visitors.R
\docType{data}
\name{STE__Year_Sex_overseas_visitors}
\alias{STE__Year_Sex_overseas_visitors}
\title{Sex by STE, Year}
\format{54 observations and 4 variables.}
\usage{
STE__Year_Sex_overseas_visitors
}
\description{
Number of overseas_visitorsSex by STE, Year
}
\keyword{datasets}
