% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_Sex_SpeaksEnglishOnly.R
\docType{data}
\name{STE__Year_Sex_SpeaksEnglishOnly}
\alias{STE__Year_Sex_SpeaksEnglishOnly}
\title{Sex, SpeaksEnglishOnly by STE, Year}
\format{108 observations and 5 variables.}
\usage{
STE__Year_Sex_SpeaksEnglishOnly
}
\description{
Number of personsSex, SpeaksEnglishOnly by STE, Year
}
\keyword{datasets}
