% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_SexemploymentToPopulation.R
\docType{data}
\name{STE__Year_SexemploymentToPopulation}
\alias{STE__Year_SexemploymentToPopulation}
\title{Sex by STE, Year}
\format{54 observations and 4 variables.}
\usage{
STE__Year_SexemploymentToPopulation
}
\description{
EmploymentSex by STE, Year #' @description ToSex by STE, Year #' @description PopulationSex by STE, Year
}
\keyword{datasets}
