% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_SexlabourForceParticipation.R
\docType{data}
\name{STE__Year_SexlabourForceParticipation}
\alias{STE__Year_SexlabourForceParticipation}
\title{Sex by STE, Year}
\format{54 observations and 4 variables.}
\usage{
STE__Year_SexlabourForceParticipation
}
\description{
LabourSex by STE, Year #' @description ForceSex by STE, Year #' @description ParticipationSex by STE, Year
}
\keyword{datasets}
