% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_Sexunemployment.R
\docType{data}
\name{STE__Year_Sexunemployment}
\alias{STE__Year_Sexunemployment}
\title{Sex by STE, Year}
\format{54 observations and 4 variables.}
\usage{
STE__Year_Sexunemployment
}
\description{
UnemploymentSex by STE, Year
}
\keyword{datasets}
