% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_SpeaksEnglishOnly.R
\docType{data}
\name{STE__Year_SpeaksEnglishOnly}
\alias{STE__Year_SpeaksEnglishOnly}
\title{SpeaksEnglishOnly by STE, Year}
\format{54 observations and 4 variables.}
\usage{
STE__Year_SpeaksEnglishOnly
}
\description{
Number of personsSpeaksEnglishOnly by STE, Year
}
\keyword{datasets}
