% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE__Year_unemployment.R
\docType{data}
\name{STE__Year_unemployment}
\alias{STE__Year_unemployment}
\title{by STE, Year}
\format{27 observations and 3 variables.}
\usage{
STE__Year_unemployment
}
\description{
Unemployment by STE, Year
}
\keyword{datasets}
